% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateSynData.R
\name{AddAllScripts}
\alias{AddAllScripts}
\title{Use the current value of \code{OutDir()} to copy scripts into all subdirectories}
\usage{
AddAllScripts()
}
\description{
Use the current value of \code{OutDir()} to copy scripts into all subdirectories
}
\keyword{internal}
