% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateSynData.R
\name{AddScript}
\alias{AddScript}
\title{Add an R script to a particular subdirectory.}
\usage{
AddScript(maxK, slice, dir.name = c("sa.sa.96", "sp.sp",
  "sa.sa.COMPOSITE", "sp.sa.COMPOSITE"))
}
\arguments{
\item{maxK}{The \code{maxK} argument for SignatureAnalyzers.}

\item{slice}{Which subdirectory to put the script into.}

\item{dir.name}{The name of the subdirectory.}
}
\description{
Add an R script to a particular subdirectory.
}
\keyword{internal}
