% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateSynData.R
\name{AssignPresentAbsentOneSample}
\alias{AssignPresentAbsentOneSample}
\title{Randomly assign present / absent to each of a set of signatures, and
keep trying until at least one is present}
\usage{
AssignPresentAbsentOneSample(prev.present, verbose = 0)
}
\arguments{
\item{prev.present}{Vector of prevalences,
each the prevalence of 1 mutational
signature.}

\item{verbose}{If > 0, cat some possibly informative messages}
}
\value{
a vector of 0s and 1s of length
\code{length(prev.present)}, and for which
\code{sum(prev.present) > 0}.
}
\description{
Randomly assign present / absent to each of a set of signatures, and
keep trying until at least one is present
}
\keyword{internal}
