% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BestSignatureAnalyzerResult.R
\name{BestSignatureAnalyzerResult}
\alias{BestSignatureAnalyzerResult}
\title{Find the best SignatureAnalyzer result over a set of output directories.}
\usage{
BestSignatureAnalyzerResult(sa.results.dir, verbose = FALSE)
}
\arguments{
\item{sa.results.dir}{The directory containing the \code{sa.run.<n>} output directories.
 For example, contents of \code{../syn.3.5.40.abst.v3/sa.sa.96/sa.results/}
 must be the directories
 \preformatted{
 sa.run.1
 sa.run.2
 sa.run.3
 ...
 }
each containing
\preformatted{
sa.output.other.data.csv
sa.output.sigs.csv
...
}

(This function only looks for subdirectories with names of the form
\code{sa.run.<n>}.)

Then
 \preformatted{
 #'  BestSignatureAnalyzerResult(
   "../syn.3.5.40.abst.v3/sa.sa.96/sa.results")
 }
 returns the path to the sa.run.<n> directory with
 the output from the best SignatureAnalyzer run, e.g.
\preformatted{"../syn.3.5.40.abst.v3/sa.sa.96/sa.results/sa.run.3/"}}

\item{verbose}{If true print informative messages to stdout.}
}
\value{
The path to the directory with the best output.
}
\description{
Find the best SignatureAnalyzer result over a set of output directories.
}
\details{
As per Jaegil, we first find the most common number
of signatures across all the runs, and then among those
runs with that number of signatures, we choose the lowest
\code{evidence} (which is the negative posterior probability).
}
