% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateMixtedTumourTypeSyntheticData.R
\name{BladderAndUV}
\alias{BladderAndUV}
\title{A simple test for CreateMixedTumorTypeSyntheticData}
\usage{
BladderAndUV()
}
\description{
A simple test for CreateMixedTumorTypeSyntheticData
}
