% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignatureAnalyzerInteraction.R
\name{CanonicalizeSAMutTypes}
\alias{CanonicalizeSAMutTypes}
\alias{CanonicalizeSAx}
\alias{CanonicalizeSADBS}
\alias{CanonicalizeSAID}
\title{Functions to canonicalize mutation type names}
\usage{
CanonicalizeSAx(feature.names)

CanonicalizeSADBS(feature.names)

CanonicalizeSAID(feature.names)
}
\value{
Mutation type names in canonical in-memory format.
}
\description{
\code{CanonicalizeSAx} For 96 and 1536 SNSs.
}
\details{
\code{CanonicalizeSADBS} For 78 DNS (doublet substitutions).

\code{CanonicalizeSAID} For indels.
}
\keyword{internal}
