% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BestSignatureAnalyzerResult.R
\name{CopyBestSignatureAnalyzerResult}
\alias{CopyBestSignatureAnalyzerResult}
\title{Find best result and make a copy of the folder}
\usage{
CopyBestSignatureAnalyzerResult(sa.results.dir, verbose = FALSE,
  overwrite = FALSE)
}
\arguments{
\item{sa.results.dir}{See \code{\link{BestSignatureAnalyzerResult}}}

\item{verbose}{See \code{\link{BestSignatureAnalyzerResult}}}

\item{overwrite}{If TRUE overwrite existing "best.run"}
}
\value{
The path of the best directory that was copied.
}
\description{
Find best result and make a copy of the folder
}
