% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Create.syn.many.types.R
\name{Create.syn.many.types}
\alias{Create.syn.many.types}
\title{Create a specific synthetic data set of 2,700 tumors}
\usage{
Create.syn.many.types()
}
\description{
Create a specific synthetic data set of 2,700 tumors
}
