% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateSynData.R
\name{CreateAndWriteCatalog}
\alias{CreateAndWriteCatalog}
\title{Create and write a mutational spectra catalog}
\usage{
CreateAndWriteCatalog(sigs, exp, dir, write.cat.fn)
}
\arguments{
\item{dir}{Directory in which to put the signatures;
NOTE: this will be a subdirectory ased on \link{\code{OutDir}}.}

\item{write.cat.fn}{Function to write catalogs \strong{or}
spectra to files.}

\item{Signatures}{to use}

\item{(Synthetic)}{exposures}
}
\value{
Invisibly, the generated catalog.
}
\description{
Create and write a mutational spectra catalog
}
\details{
Creats a file with the catalog \code{syn.data.csv}
 and writes \code{sigs} to \code{input.sigs.csv}.
}
