% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Create.syn.30.random.R
\name{CreateExposuresNums}
\alias{CreateExposuresNums}
\title{Create \code{num.exposures} signature counts
from a normal distribution with \code{mean} and
\code{sd}.}
\usage{
CreateExposuresNums(num.exposures, mean, sd, total.num.sigs)
}
\arguments{
\item{num.exposures}{Number of exposures to create.}

\item{mean}{Mean of distribution to draw from.}

\item{sd}{Standard deviation of distribution to draw from.}

\item{total.num.sigs}{Number of signatures in the "universe".}
}
\value{
A numeric vector, each element of which the number
of signatures in the corresponding tumor (which still remains
to be created.)
}
\description{
Discard tumors with signature count = 0 or
signature count > \code{total.num.sigs}.
}
\keyword{internal}
