% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Create.syn.30.random.R
\name{CreateMeanAndStdevForSigs}
\alias{CreateMeanAndStdevForSigs}
\title{Create means and standard deviations of log10 mutation counts for synthetic
signatures.}
\usage{
CreateMeanAndStdevForSigs(num.sigs, target.mut.mean, target.mut.sd,
  sig.names)
}
\arguments{
\item{num.sigs}{Number of signatures for which means and standard deviations
are needed.}

\item{target.mut.mean}{Target number of mean of log10 of number
of mutations per tumor due to one signature.}

\item{target.mut.sd}{Target of standard deviation of  the the log10 of the
number of mutations per tumor due to one signature.}

\item{sig.names}{Names for the output vectors.}
}
\value{
A list with the needed synthetic means
and standard deviations.
}
\description{
Create means and standard deviations of log10 mutation counts for synthetic
signatures.
}
\keyword{internal}
