% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateMixtedTumourTypeSyntheticData.R
\name{CreateMix}
\alias{CreateMix}
\title{Create a specific synthetic data set of 2,700 tumors}
\usage{
CreateMix()
}
\description{
Create a specific synthetic data set of 2,700 tumors
}
