% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Create.syn.many.types.R
\name{CreateMixedTumorTypeSyntheticData}
\alias{CreateMixedTumorTypeSyntheticData}
\title{Create a full SignatureAnalyzer / SigProfiler test data set for a
set of various tumor types.}
\usage{
CreateMixedTumorTypeSyntheticData(top.level.dir, cancer.type.strings,
  num.syn.tumors, overwrite = FALSE, verbose = 0)
}
\arguments{
\item{top.level.dir}{Path to top level of directory structure to be created.}

\item{cancer.type.strings}{Search the PCAWG data for tumors matching
these strings. Each string should identify one tumor type, for
some definition of tumor type. Probably the tumors in each type
should be non-overlapping, but the code does not enforce this and
does not care.}

\item{num.syn.tumors}{Number of synthetic tumors to create
for each cancer type.}

\item{overwrite}{If TRUE, overwrite existing directories / files.}

\item{verbose}{If > 0, cat various messages.}
}
\description{
Create a full SignatureAnalyzer / SigProfiler test data set for a
set of various tumor types.
}
