% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiModalMuSigInteraction.R
\name{CreateMultiModalMuSigOutput}
\alias{CreateMultiModalMuSigOutput}
\title{Prepare input file for MultiModalMuSig from a
MultiModalMuSig formatted catalog file.}
\usage{
CreateMultiModalMuSigOutput(catalog = "ground.truth.syn.catalog.csv",
  read.catalog.function = NULL, out.dir = "MultiModalMuSig.results",
  overwrite = FALSE)
}
\arguments{
\item{catalog}{a catalog in ICAMS format. It can be
a .csv file, or a matrix or data.frame.
Usually, it refers to \code{"ground.truth.syn.catalog.csv"}.}

\item{read.catalog.function}{Function taking a file path as
its only argument and returning a catalog as a numeric matrix.}

\item{out.dir}{Directory that will be created for the output;
abort if it already exits.  Log files will be in
\code{paste0(out.dir, "/tmp")}.}

\item{overwrite}{If TRUE, overwrite existing output}
}
\value{
\code{invisible(catMatrix)},
original catalog in MultiModalMuSig format
}
\description{
Prepare input file for MultiModalMuSig from a
MultiModalMuSig formatted catalog file.
}
\details{
Creates folder named \code{MultiModalMuSig.results} containing catalogs
in MultiModalMuSig-formatted catalogs: Rows are signatures;
the first column is the name of the mutation type, while the remaining
columns are samples (tumors).
These MM-formated catalogs will the input when running MultiModalMuSig program
later on Julia platform.
}
