% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Create.syn.30.random.R
\name{CreateRandomMutSigProfiles}
\alias{CreateRandomMutSigProfiles}
\title{Create a matrix of "random" signature profiles.}
\usage{
CreateRandomMutSigProfiles(row.headers, num.signatures, sig.name.prefix)
}
\arguments{
\item{row.headers}{One of the \code{\link{ICAMS}} package variable such as
\code{catalog.row.order[["SBS96"]]}.}

\item{num.signatures}{Number of signatures to create.}

\item{sig.name.prefix}{The signatures will be named \code{<sig.name.prefix>1},
\code{<sig.name.prefix>2}, etc.}
}
\value{
A \code{num.signatures}-column
  matrix with rownames \code{row.headers}.
}
\description{
Create a matrix of "random" signature profiles.
}
\keyword{internal}
