% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateTotalRandomSynData2.R
\name{CreateSAAndSPSynCatalogs}
\alias{CreateSAAndSPSynCatalogs}
\title{Create a full SignatureAnalyzer / SigProfiler test data set.}
\usage{
CreateSAAndSPSynCatalogs(top.level.dir, num.syn.tumors,
  overwrite = FALSE)
}
\arguments{
\item{top.level.dir}{Path to top level of directory structure to be created.}

\item{overwrite}{If TRUE, overwrite existing directories / files.}

\item{num.sig.tumors}{Number of synthetic tumors to create.}
}
\description{
Create a full SignatureAnalyzer / SigProfiler test data set.
}
