% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Create.syn.SBS1.5.correlated.R
\name{CreateSBS1SBS5CorrelatedSyntheticDataDemo}
\alias{CreateSBS1SBS5CorrelatedSyntheticDataDemo}
\title{Function to generate 20 SBS1-SBS5-correlated Synthetic datasets used in testing.}
\usage{
CreateSBS1SBS5CorrelatedSyntheticDataDemo(top.level.dir = "./",
  overwrite = FALSE)
}
\arguments{
\item{top.level.dir}{Top-level-folder to place 20 spectra
datasets generated by this function.
Default: ./ (Current working directory)}

\item{overwrite}{Whether to overwrite
(Default: FALSE)}
}
\description{
This function is a wrapper around \link{CreateSBS1SBS5CorrelatedSyntheticData}.
It will use the default parameters to repeat the results.
}
\details{
This function will generate 20 datasets, each with files listed below:

ground.truth.syn.catalog.csv: Generated tumor spectra in
ICAMS SBS96 CSV format.

ground.truth.syn.exposures.csv: Mutation burdens of SBS1 and
SBS5 in generated tumor spectra in ICAMS CSV format.

ground.truth.syn.sigs.csv: Ground-truth SBS1 and SBS5
signatures in ICAMS SBS96 CSV format.

parameters.txt: Parameters used to generate the exposures
and tumor spectra.

scatterplot.pdf: scatterplot illustrating correlation of
exposures of two signatures in generated spectra

seedInUse.txt, RNGInUse.txt: seed and Random Number Generator
used in generation. (For better reproducibility)

sessionInfo.txt: information related to R versions, platforms,
loaded or imported packages, etc. (For better reproducibility)
}
