% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateSynData.R
\name{CreateSynCatalogs}
\alias{CreateSynCatalogs}
\title{Generate synthetic spectra catalogs given
signature profiles and synthetic exposures.}
\usage{
CreateSynCatalogs(signatures, exposures, sample.id.suffix = NULL)
}
\arguments{
\item{signatures}{The signature profiles.}

\item{exposures}{The synthetic exposures.}

\item{sample.id.suffix}{A string for adding a suffix to
sample ID. For example, if sample.id.suffix is "abc",
then SomeCancerType::s1.33 is changed to
SomeCancerType::s1-abc.33. Actually, this just replaces
the first "." in the sample id with "-" concatenated
to sample.id.suffix. TODO(Steve): probably drop this}
}
\value{
A list of three elements that comprise the
synthetic data: \enumerate{
 \item \code{ground.truth.catalog}: Spectra catalog for
 the software input.
 \item \code{ground.truth.signatures}: Signatures active
 in \code{ground.truth.catalog}.
 \item \code{ground.truth.exposures}: Exposures of \code{ground.truth.signatures}
 in \code{ground.truth.catalog}.
}
}
\description{
Generate synthetic spectra catalogs given
signature profiles and synthetic exposures.
}
