% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Diff4SynDataSets.R
\name{Diff4SynDataSets}
\alias{Diff4SynDataSets}
\title{diff new directory / files against regression data for testing.}
\usage{
Diff4SynDataSets(dirname)
}
\arguments{
\item{dirname}{the root name of the directories to diff.}
}
\value{
The return value of diff command.
}
\description{
diff new directory / files against regression data for testing.
}
