% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SigSimilarity.R
\name{FindNearestSig}
\alias{FindNearestSig}
\title{Find the signature in other.sigs that is nearest (by cosine similarity)
to query.sig.}
\usage{
FindNearestSig(query.sig, other.sigs)
}
\arguments{
\item{query.sig}{A single signature}

\item{other.sigs}{Matrix with each column being one signature}
}
\value{
The maximum similarity between \code{query.sig} and any signature in
\code{other.sigs}
}
\description{
Find the signature in other.sigs that is nearest (by cosine similarity)
to query.sig.
}
