% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignatureAnalyzerInteraction.R
\name{FixSASigNames}
\alias{FixSASigNames}
\title{Standardize SignatureAnalyzer signature names}
\usage{
FixSASigNames(sig.names)
}
\arguments{
\item{sig.names}{Vector of signature names}
}
\value{
Vector of signatures names with "_SNV" removed.
}
\description{
For example, change \code{BI_COMPOSITE_SNV_SBS83_P}
to \code{BI_COMPOSITE_SBS83_P}
}
\details{
This is necessary because
for COMPOSITE signatures we rbind coordinated
"SNV", "DNP", and "INDEL" signatures.
}
