% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignatureAnalyzer_interaction.R
\name{fwriteDataFrame}
\alias{fwriteDataFrame}
\title{Probably not needed}
\usage{
fwriteDataFrame(df, file, rowname.name = "mutation.type")
}
\arguments{
\item{df}{A data.frame}

\item{file}{File to write to}

\item{rowname.name}{Name of first column in file}
}
\description{
Probably not needed
}
