% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateSynData.R
\name{generate.synthetic.exposures}
\alias{generate.synthetic.exposures}
\title{Create synthetic exposures based given parameters}
\usage{
generate.synthetic.exposures(sig.params, num.samples = 10,
  name = "synthetic")
}
\arguments{
\item{sig.params}{Parameters from \code{synsig.parameters.from.attribution}}

\item{num.samples}{Number of samples to generate}

\item{name}{Prefix for sample identifiers in the simulated dataset}
}
\value{
A matrix with the rows being each signature and the columns being
generated samples. Each entry is the count of mutations due to one
signature in one sample.
}
\description{
Create synthetic exposures based given parameters
}
