% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateSynData.R
\name{GenerateSynExposureOneSample}
\alias{GenerateSynExposureOneSample}
\title{TODO(steve) trace this to be sure we understand what it does.}
\usage{
GenerateSynExposureOneSample(tumor, sig.interest, burden.per.sig,
  sd.per.sig)
}
\arguments{
\item{tumor}{TODO}

\item{sig.interest}{TODO}

\item{burden.per.sig}{TODO}

\item{sd.per.sig}{TODO}
}
\description{
TODO(steve) trace this to be sure we understand what it does.
}
\details{
??Determine the intensity of each
mutational signature in a tumor, returning mutations per mb
using the mean mutation burden per signature and the std dev
}
\keyword{internal}
