% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateSynData.R
\name{GenerateSynFromReal}
\alias{GenerateSynFromReal}
\title{Generate synthetic exposures from real exposures.}
\usage{
GenerateSynFromReal(real.exp, num.syn.tumors, file.prefix,
  sample.id.prefix)
}
\arguments{
\item{real.exp}{The actual (real) exposures upon which to base
the parameters and synthetic exposures.}

\item{num.syn.tumors}{Generate this number of synthetic tumors.}

\item{file.prefix}{Prepend this to output filenames
to indicate the organization of the data.}

\item{sample.id.prefix}{Prefix for sample identifiers for the
synthetic samples.}
}
\value{
A list with elements:
\enumerate{
 \item \code{parms} The parameters inferred from \code{real.exp}.
 \item \code{syn.exp} The synthetic exposures generated from \code{parms}.
}
}
\description{
Checkpoints the parameters and the synthetic
exposures to files. It also checks that the parameters
inferred from the synthetic data approximate those
inferred from \code{real.exp}.
}
