% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Create.syn.SBS1.5.correlated.R
\name{GenSBS1SBS5ExposureOneTumor}
\alias{GenSBS1SBS5ExposureOneTumor}
\title{Generate correlated exposures for one tumor

Function to generate exposure of two correlated signatures
(Example: SBS1 and SBS5) for ONE synthetic tumor.}
\usage{
GenSBS1SBS5ExposureOneTumor(tumor.name = "TwoCorreSigsGen::1",
  main.signature = "SBS5", correlated.signature = "SBS1",
  main.mean.log = 2.5, main.stdev.log = 0.25,
  correlated.stdev.log = 0.25, slope.linear = 1,
  main.signature.lower.thres = 50,
  correlated.signature.lower.thres = 30,
  min.main.to.correlated.ratio.linear = 1/3,
  max.main.to.correlated.ratio.linear = Inf)
}
\arguments{
\item{tumor.name}{Name of synthetic tumor you want to generate.
(default: "TwoCorreSigsGen::1")}

\item{main.signature}{Name of a signature with smaller variance
in the log10 space. (default: "SBS5")}

\item{correlated.signature}{Name of a signature with larger variance
in the log10 space. (default: "SBS1")}

\item{main.mean.log}{Mean of log10(mutation burden of \code{main.signature})}

\item{main.stdev.log}{Standard deviation of log10(mutation burden
of \code{main.signature})}

\item{correlated.stdev.log}{Contribute to part of the standard deviation of
log10(mutation burden of correlated.signature). In this script, the s.d. of
log10(mutation burden of correlated.signature)
= main.stdev.log + correlated.stdev.log}

\item{slope.linear}{Average ratio of mutation burden of \code{correlated.signature}
over mutation burden of \code{main.signature}}

\item{main.signature.lower.thres}{Minimum mutation burden (number of mutations)
induced by \code{main.signature} in each tumor.}

\item{correlated.signature.lower.thres}{Minimum mutation burden (number of mutations)
induced by \code{correlated.signature} in each tumor.}

\item{min.main.to.correlated.ratio.linear}{Minimum ratio of
\code{main.signature} over mutation burden of
\code{correlated.signature} in each tumor.}

\item{max.main.to.correlated.ratio.linear}{Maximum ratio of
\code{main.signature} over mutation burden of
\code{correlated.signature} in each tumor.}
}
\description{
Generate correlated exposures for one tumor

Function to generate exposure of two correlated signatures
(Example: SBS1 and SBS5) for ONE synthetic tumor.
}
\section{Warning Warning 1}{
 forcing dataset to have high Pearson's R^2,
while inputting large main.stdev.log value will cause
the program to reject unqualified datasets and regenerate
new datasets for HOURS OR EVEN DAYS!

Warning 2: For replication of paperresults, only specify
parameters as indicated by supplementary information!
}

\keyword{internal}
