% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateSynData.R
\name{GenSynCatalogs}
\alias{GenSynCatalogs}
\title{Generate synthetic spectra catalogs given
signature profiles and synthetic exposures.}
\usage{
GenSynCatalogs(signatures, exposures, sample.id.suffix = NULL)
}
\arguments{
\item{signatures}{The signature profiles.}

\item{exposures}{The synthetic exposures.}

\item{sample.id.suffix}{A string for adding a suffix to
sample ID. For example, if sample.id.suffix is "abc",
then SomeCancerType::s1.33 is changed to
SomeCancerType::s1-abc.33. Actually, this just replaces
the first "." in the sample id with "-" concatenated
to sample.id.suffix. TODO(Steve): probably drop this}
}
\value{
A lsit of three elements TODO(Steve): FINISH Spectra catalog as a numeric matrix.
}
\description{
Generate synthetic spectra catalogs given
signature profiles and synthetic exposures.
}
