% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiModalMuSigInteraction.R
\name{ICAMSCatalog2MM}
\alias{ICAMSCatalog2MM}
\title{Convert Catalogs from ICAMS format to MM format}
\usage{
ICAMSCatalog2MM(catalog)
}
\arguments{
\item{catalog}{A catalog matrix in ICAMS format. (SNS/DNS/ID)}
}
\value{
a catalog matrix in MultiModalMuSig format.
}
\description{
Convert Catalogs from ICAMS format to MM format
}
