% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RundeconstructSigsAttributionOnly.R
\name{InstalldeconstructSigs}
\alias{InstalldeconstructSigs}
\title{Install deconstructSigs from CRAN}
\usage{
InstalldeconstructSigs()
}
\description{
Install deconstructSigs from CRAN
}
