% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunMutationalPatterns.R
\name{InstallMutationalPatterns}
\alias{InstallMutationalPatterns}
\title{Install MutationalPatterns from Bioconductor,
and its dependent package, NMF.}
\usage{
InstallMutationalPatterns()
}
\description{
Install MutationalPatterns from Bioconductor,
and its dependent package, NMF.
}
\keyword{internal}
