% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Runsigfit.R
\name{Installsigfit}
\alias{Installsigfit}
\title{Install sigfit from GitHub,
and its dependent package, rstan.}
\usage{
Installsigfit()
}
\description{
Install sigfit from GitHub,
and its dependent package, rstan.
}
\keyword{internal}
