% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunSignatureEstimationAttributionOnly.R
\name{InstallSignatureEstimation}
\alias{InstallSignatureEstimation}
\title{Install SignatureEstimation package from URL source.}
\usage{
InstallSignatureEstimation()
}
\description{
Install SignatureEstimation package from URL source.
}
\keyword{internal}
