% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SigSimilarity.R
\name{Match1Sig}
\alias{Match1Sig}
\title{Find the signature in other.sigs that is nearest (by cosine similarity)
to query.sig.}
\usage{
Match1Sig(query.sig, other.sigs)
}
\arguments{
\item{query.sig}{A single signature}

\item{other.sigs}{Matrix with each column being one signature}
}
\value{
The maximum similarity between \code{query.sig} and any signature in
\code{other.sigs}
}
\description{
Find the signature in other.sigs that is nearest (by cosine similarity)
to query.sig.
}
\seealso{
Other signature matching functions: \code{\link{MatchSigs1Direction}},
  \code{\link{MatchSigs2Directions}},
  \code{\link{MatchSigsThenWriteAndPlot}}
}
\concept{signature matching functions}
