% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SigSimilarity.R
\name{MatchSigs}
\alias{MatchSigs}
\title{Find the closest match in \code{other.sigs} for each signature in \code{query.sigs}}
\usage{
MatchSigs(query.sigs, other.sigs)
}
\arguments{
\item{query.sigs}{A signature matrix; signatures for which to find the
closest match in other.sigs. The colnames are used
  as the identifiers of the signatures.}

\item{other.sigs}{A signature matrix; find the closest matches to
a signature in this matrix.
The colnames are used as the identifiers of the signatures.}
}
\value{
A list with one element for each signature in query.sigs. The names
  of the list elements are the colnames of query.sigs. Each list element
  is a vector of length 1, and the name of the vector element is
  the name of the closest matching signature in other.sigs, and the value
  is the cosine similarily between the given signature in query.sigs and
  the matching signatute in other.sigs.
}
\description{
Find the closest match in \code{other.sigs} for each signature in \code{query.sigs}
}
