% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SigSimilarity.R
\name{MatchSigs2Directions}
\alias{MatchSigs2Directions}
\title{Calculate bidirectional closest similarities bewteen two sets of signatures
and the average of the similarities.}
\usage{
MatchSigs2Directions(sigs1, sigs2)
}
\arguments{
\item{sigs1}{Matrix of signatures; colnames are used as signature
identifiers, and the colnams in \code{sigs1} should be distinguisable from
those in \code{sigs2}.}

\item{sigs2}{Matrix of signatures; colnames are used as signature identifiers.}
}
\value{
A list with the elements:

\code{avg}: the average of the cosine similarities between each signature in
\code{sigs1} and its closest match in \code{sigs2} and the closest match
between each signature in \code{sigs2} and its closest match in \code{sigs1}.

\code{match1}: a data frame with rownames being signature identifiers from
\code{sigs1}, the signature identifier of the closest match in \code{sigs1}
in the 1st column, and the cosine similarity between them in the 2nd column.

\code{match2}: a data frame with the rownames being signature identifiers
from \code{sigs2}, the signature identifier of the closest match in
\code{sigs1} in the 1st column, and the cosine similarity between them in the
2nd column.

\code{match1} and \code{match2} might not have the same number of rows.
}
\description{
Calculate bidirectional closest similarities bewteen two sets of signatures
and the average of the similarities.
}
\seealso{
Other signature matching functions: \code{\link{Match1Sig}},
  \code{\link{MatchSigs1Direction}},
  \code{\link{MatchSigsThenWriteAndPlot}},
  \code{\link{TestMatchSigs2Directions}}
}
\concept{signature matching functions}
