% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssessExtracted.R
\name{MatchSigsAndRelabel}
\alias{MatchSigsAndRelabel}
\title{Run \code{MatchSigs2Directions}, then
plot its results and write them as .csv files.}
\usage{
MatchSigsAndRelabel(ex.sigs, gt.sigs, exposure)
}
\arguments{
\item{ex.sigs}{Newly extracted signatures to be compared to gt.sigs}

\item{gt.sigs}{"Ground truth" signatures.}

\item{exposure}{"Ground truth" exposures used generate the
synthetic data from which \code{ex.sigs} were extracted.}
}
\value{
A list with the elements \code{avg}, \code{match1},
\code{match2} as for \code{SigSetSimilarity}, with \code{match1}
 being matches for the the extracted signatures (\code{ex.sigs})
 and \code{match2} being the
 matches for the ground truth signatures (\code{gt.sigs}). The return list
 also echos the input arguments \code{ex.sigs} and \code{gt.sigs}.
}
\description{
Run \code{MatchSigs2Directions}, then
plot its results and write them as .csv files.
}
\seealso{
Other signature matching functions: \code{\link{Match1Sig}},
  \code{\link{MatchSigs1Direction}},
  \code{\link{MatchSigs2Directions}}
}
\concept{signature matching functions}
