% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SigSimilarity.R
\name{MatchSigsThenWriteAndPlot}
\alias{MatchSigsThenWriteAndPlot}
\title{Run \code{MatchSigs2Directions}, then
plot its results and write them as .csv files.}
\usage{
MatchSigsThenWriteAndPlot(ex.sigs, gt.sigs, plot.fn, file.prefix = "")
}
\arguments{
\item{ex.sigs}{Newly extracted signatures to be compared to gt.sigs}

\item{gt.sigs}{"Ground truth" signatures}

\item{plot.fn}{Function to plot signatures. For "typical" 96-channel}

\item{file.prefix}{A prefix string for output files, can be used to put all}
}
\value{
A list with the elments \code{avg}, \code{match1},
\code{match2} as for \code{SigSetSimilarity}, with \code{match1}
 being matches for the the extracted signatures (\code{ex.sigs})
 and \code{match2} being the
 matches for the ground truth signatures (\code{gt.sigs}). The return list
 also echos the input arguments \code{ex.sigs} and \code{gt.sigs}.
}
\description{
Run \code{MatchSigs2Directions}, then
plot its results and write them as .csv files.
}
\details{
Uses \code{plot.fn} to
plot the extracted signatures with identifiers showing the
nearest ground-truth signatures in a location defined by
\code{file.prefix}. Saves the match tables to
.csv file with suffixes \code{match.extracted.to.gt.csv}
and \code{match.gt.to.extracted.csv}.
}
\seealso{
Other signature matching functions: \code{\link{Match1Sig}},
  \code{\link{MatchSigs1Direction}},
  \code{\link{MatchSigs2Directions}},
  \code{\link{TestMatchSigs2Directions}}
}
\concept{signature matching functions}
