% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateSynData.R
\name{Merge2Exposures}
\alias{Merge2Exposures}
\title{Merge 2 exposure matrices}
\usage{
Merge2Exposures(exp1, exp2)
}
\arguments{
\item{exp1}{An exposure matrix}

\item{exp2}{An exposure matrix}
}
\value{
The column-wise merge of the two input matrices as
with all rownames from either matrix preserved and
corresponding entries filled with 0s.
}
\description{
Merge 2 exposure matrices
}
\keyword{internal}
