% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateSynData.R
\name{MergeExposures}
\alias{MergeExposures}
\title{Merge all exposure matrices in a list of matrices}
\usage{
MergeExposures(list.of.exposures)
}
\arguments{
\item{list.of.exposures}{A list of exposure matrices}
}
\value{
The column-wise merge of all the input matrices
with all rownames from all matrices preserved and
corresponding entries filled with 0s.
}
\description{
Merge all exposure matrices in a list of matrices
}
