% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiModalMuSigInteraction.R
\name{MMCatalog2ICAMS}
\alias{MMCatalog2ICAMS}
\title{Convert Catalogs from MM format to ICAMS format}
\usage{
MMCatalog2ICAMS(cat)
}
\arguments{
\item{cat}{Input catalog, can be a tab-delimited file
or matrix in MultiModalMuSig format.}
}
\value{
a catalog matrix in ICAMS format.
}
\description{
Convert Catalogs from MM format to ICAMS format
}
