% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateSynData.R
\name{NewCreateAndWriteCatalog}
\alias{NewCreateAndWriteCatalog}
\title{Create and write a mutational spectra catalog}
\usage{
NewCreateAndWriteCatalog(sigs, exp, dir, extra.file.suffix = "",
  overwrite = FALSE)
}
\arguments{
\item{sigs}{Signatures to use.}

\item{exp}{(Synthetic) exposures.}

\item{dir}{Directory in which to put the signatures;
NOTE: this will be a subdirectory based on \code{\link{OutDir}}.}

\item{extra.file.suffix}{Extra string to put before ".csv".}

\item{overwrite}{If TRUE, overwrite existing directory; useful for
debugging / testing.}
}
\value{
Invisibly, the generated catalog.
}
\description{
Create and write a mutational spectra catalog
}
\details{
Create a file with the catalog \code{syn.data.csv}
 and writes \code{sigs} to \code{input.sigs.csv}.
}
