% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SigSimilarity.R
\name{NumFromId}
\alias{NumFromId}
\title{Get the numerical parts of signature ids}
\usage{
NumFromId(s)
}
\arguments{
\item{s}{A character vector}
}
\value{
A vector, each element of which is the integer
corresponding to the first string of digits of an element of s
}
\description{
Get the numerical parts of signature ids
}
