% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssessExtracted.R
\name{OutDir}
\alias{OutDir}
\title{Create file names in a given directory}
\usage{
OutDir(file.name)
}
\arguments{
\item{file.name}{The name of the that will be
prefixed by \code{out.data.dir}.}
}
\value{
\code{file.name} prefixed by \code{out.data.dir}.
}
\description{
The directory is provided by \code{out.data.dir},
which must be set by the user.
}
