% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Create.syn.SBS1.5.correlated.R
\name{plot.correlation.scatterplot}
\alias{plot.correlation.scatterplot}
\title{Plot scatter plot for correlation between exposures of two signatures

Plot scatter plot for correlation between exposures of two signatures,
SBS1 and SBS5 in this study.

\code{plot.correlation.scatterplot.for.exposures} is a
function to plot the correlation between two vectors,
\code{x} and \code{y}. These vectors are expected to be
exposures of two signatures.
It will draw a scatterplot, and it will also print information
onto the plot, including correlation between \code{x} and \code{y},
mean and stdev of \code{x} and \code{y}, etc.

\code{plot.correlation.scatterplot.for.exposures}
is a wrapper around \code{plot.correlation.scatterplot}.
It lets \code{exposure.counts <-} the exposure matrix,
and will draw a scatterplot for exposures of two signatures.}
\usage{
\method{plot}{correlation.scatterplot}(x, y, xlab = NULL, ylab = NULL,
  main = NULL, optional.remarks = "", ...)
}
\arguments{
\item{x}{vector of exposures of \code{main.signature}
(SBS5 in the paper).
The exposures of \code{main.siganture} will be aligned onto x axis.}

\item{y}{vector of exposures of \code{correlated.signature}
(SBS1 in the paper).
The exposures of \code{correlated.siganture} will be aligned onto y axis.}

\item{xlab}{Label below x axis.}

\item{ylab}{Label below y axis.}

\item{main}{Title on the scatterplot.
Default: NULL}

\item{optional.remarks}{Remarks added below the title.}

\item{pdf.filename}{Name of the PDF to contain the scatterplots.}

\item{main.signature}{Name of a signature with smaller variance
in the log10 space. (Default: "SBS5")}

\item{correlated.signature}{Name of a signature with larger variance
in the log10 space. (Default: "SBS1")}

\item{slope.linear}{Average ratio of mutation burden of \code{correlated.signature}
over mutation burden of \code{main.signature}}

\item{exposure.counts}{Data.frame or matrix storing exposures of two signatures.
The exposure.counts object is usually obtained from \code{SynSig::ReadExposure()}.}

\item{xlim, ylim}{numeric vectors of length 2,
giving the x and y coordinates ranges.
Default: c(0,4)}
}
\description{
Plot scatter plot for correlation between exposures of two signatures

Plot scatter plot for correlation between exposures of two signatures,
SBS1 and SBS5 in this study.

\code{plot.correlation.scatterplot.for.exposures} is a
function to plot the correlation between two vectors,
\code{x} and \code{y}. These vectors are expected to be
exposures of two signatures.
It will draw a scatterplot, and it will also print information
onto the plot, including correlation between \code{x} and \code{y},
mean and stdev of \code{x} and \code{y}, etc.

\code{plot.correlation.scatterplot.for.exposures}
is a wrapper around \code{plot.correlation.scatterplot}.
It lets \code{exposure.counts <-} the exposure matrix,
and will draw a scatterplot for exposures of two signatures.
}
