% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignatureAnalyzerInteraction.R
\name{Plot96PartOfComposite}
\alias{Plot96PartOfComposite}
\title{Plot the SBS96 part of a SignatureAnalyzer COMPOSITE signature or catalog}
\usage{
Plot96PartOfComposite(catalog, name, type = "density")
}
\arguments{
\item{catalog}{Catalog or signature matrix}

\item{name}{Name of file to print to.}

\item{type}{See \code{\link[ICAMS]{PlotCatSNS96ToPdf}}.}
}
\description{
TODO(Steve): Deal with Collapse1536to96
}
