% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotCatalogs.R
\name{PlotCatCOMPOSITE}
\alias{PlotCatCOMPOSITE}
\title{Plot the a SignatureAnalyzer COMPOSITE signature or catalog into separate pdfs}
\usage{
PlotCatCOMPOSITE(catalog, filename.header, type, id = colnames(catalog))
}
\arguments{
\item{catalog}{Catalog or signature matrix}

\item{filename.header}{Contain path and the beginning part of the file name.
The name of the pdf files will be:
\code{filename.header}.SNS.96.pdf
\code{filename.header}.SNS.1536.pdf
\code{filename.header}.DNS.78.pdf
\code{filename.header}.ID.83.pdf}

\item{type}{See \code{\link[ICAMS]{PlotCatalogToPdf}}.}

\item{id}{A vector containing the identifiers of the samples
or signatures in \code{catalog}.}
}
\description{
Plot the a SignatureAnalyzer COMPOSITE signature or catalog into separate pdfs
}
