% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Create.syn.SBS1.5.correlated.R
\name{PlotCorrelationScatterplot}
\alias{PlotCorrelationScatterplot}
\title{Plot scatter plot for correlation between two vectors.



\code{PlotCorrelationScatterplot} is a wrapper around \code{graphics::plot()},
and a function to plot the correlation between two vectors,
\code{x} and \code{y}. These vectors are expected to be
exposures of two signatures.

It will draw a scatterplot, and it will also print information
onto the plot, including correlation between \code{x} and \code{y},
mean and stdev of \code{x} and \code{y}, etc.}
\usage{
PlotCorrelationScatterplot(x, y, xlab = NULL, ylab = NULL,
  main = NULL, optional.remarks = "", ...)
}
\arguments{
\item{x}{vector of exposures of \code{main.signature}
(SBS5 in the paper).
The exposures of \code{main.siganture} will be aligned onto x axis.}

\item{y}{vector of exposures of \code{correlated.signature}
(SBS1 in the paper).
The exposures of \code{correlated.siganture} will be aligned onto y axis.}

\item{xlab}{Label below x axis.}

\item{ylab}{Label below y axis.}

\item{main}{Title on the scatterplot.
Default: NULL}

\item{optional.remarks}{Remarks added below the title.}

\item{...}{Other parameters provided to the function \code{graphics::plot()}.}
}
\description{
Plot scatter plot for correlation between two vectors.



\code{PlotCorrelationScatterplot} is a wrapper around \code{graphics::plot()},
and a function to plot the correlation between two vectors,
\code{x} and \code{y}. These vectors are expected to be
exposures of two signatures.

It will draw a scatterplot, and it will also print information
onto the plot, including correlation between \code{x} and \code{y},
mean and stdev of \code{x} and \code{y}, etc.
}
