% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Create.syn.SBS1.5.correlated.R
\name{PlotCorrelationScatterplotForExposures}
\alias{PlotCorrelationScatterplotForExposures}
\title{Plot scatter plot for correlation between exposures of two signatures

Plot scatter plot for correlation between exposures of two signatures,
SBS1 and SBS5 in this study.

\code{PlotCorrelationScatterplotForExposures}
is a wrapper around \code{\link{PlotCorrelationScatterplot}}.
It lets \code{exposure.counts <-} the exposure matrix,
and will draw a scatterplot for exposures of two signatures.}
\usage{
PlotCorrelationScatterplotForExposures(pdf.filename,
  main.signature = "SBS5", correlated.signature = "SBS1", slope.linear,
  exposure.counts, xlim = c(0, 4), ylim = c(0, 4), ...)
}
\arguments{
\item{pdf.filename}{Name of the PDF to contain the scatterplots.}

\item{main.signature}{Name of a signature with smaller variance
in the log10 space. (Default: "SBS5")}

\item{correlated.signature}{Name of a signature with larger variance
in the log10 space. (Default: "SBS1")}

\item{slope.linear}{Average ratio of mutation burden of \code{correlated.signature}
over mutation burden of \code{main.signature}}

\item{exposure.counts}{Data.frame or matrix storing exposures of two signatures.
The exposure.counts object is usually obtained from \code{SynSig::ReadExposure()}.}

\item{xlim, ylim}{numeric vectors of length 2,
giving the x and y coordinates ranges.
Default: c(0,4)}

\item{...}{Other parameters provided to the function \code{graphics::plot()}.}

\item{main}{Title on the scatterplot.
Default: NULL}
}
\description{
Plot scatter plot for correlation between exposures of two signatures

Plot scatter plot for correlation between exposures of two signatures,
SBS1 and SBS5 in this study.

\code{PlotCorrelationScatterplotForExposures}
is a wrapper around \code{\link{PlotCorrelationScatterplot}}.
It lets \code{exposure.counts <-} the exposure matrix,
and will draw a scatterplot for exposures of two signatures.
}
