% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SynSigFunctions.v0.7.R
\name{present.sigs}
\alias{present.sigs}
\title{Decide which signatures are present in the catalogs of synthetic tumors.}
\usage{
present.sigs(num.tumors, prev.present, sigs)
}
\arguments{
\item{num.tumors}{Number of tumors to generate}

\item{prev.present}{Vector of prevalences, each the prevalence of 1 mutationa
signature}

\item{sigs}{List(?) maybe vector(?) of signature names (?)}
}
\description{
Decide which signatures are present in the catalogs of synthetic tumors.
}
\details{
If a tumor ends up with no signature assigned,
signature 1 is added as the only signature. TODO:(steve)
needs redesign how to handle 0 sig assigned cases
}
