% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssessExtracted.R
\name{ReadAndAnalyzeExposures}
\alias{ReadAndAnalyzeExposures}
\title{Assess how well attributed exposures match input exposures

We assume that in many cases attribution programs will be run
outside of R on file inputs and will generate fill outputs.}
\usage{
ReadAndAnalyzeExposures(extracted.sigs, ground.truth.sigs,
  attributed.exposures, ground.truth.exposures,
  read.extracted.sigs.fn = NULL, read.ground.truth.sigs.fn)
}
\arguments{
\item{extracted.sigs}{Path to file containing the extracted signature profiles.}

\item{ground.truth.sigs}{File containing signature profiles from which the
synthetic data were generated.}

\item{attributed.exposures}{File containing mutation counts (exposures)
of synthetic tumors which are attributed to extracted or input signatures.}

\item{ground.truth.exposures}{File containing the exposures from which
the synthetic catalogs were generated.  This file is used to restrict
assessment of signature exposures to only those signatures in
\code{ground.truth.sigs} that were actually represented in the exposures.}

\item{read.extracted.sigs.fn}{Function to read the extracted signatures
into the appropriate standard internal representation. If NULL then
\code{read.ground.truth.sigs.fn} is used.}

\item{read.ground.truth.sigs.fn}{Function to read the ground truth
signatures into the appropriate standard internal representation.}
}
\value{
A \code{\link{data.frame}} recording:

\code{Ground.truth.exposure}: sum of ground truth exposures of
all tumors to all ground-truth signatures.

\code{Attributed.exposure}: sum of attributed exposures of
all tumors to all ground-truth signatures.
Here, attributed exposure of a tumor to a ground-truth
signature equals to the sum of the exposures of this tumor
to all extracted signatures which are most similar to
a ground-truth signature.
If there is no extracted signature resembling an ground-truth
signature, the attributed exposure of this ground-truth
signature will be \code{0}.

\code{Absolute.difference}: sum of absolute difference between
ground-truth exposure and attributed exposure of all tumors
to all ground-truth signatures.
}
\description{
Assess how well attributed exposures match input exposures

We assume that in many cases attribution programs will be run
outside of R on file inputs and will generate fill outputs.
}
\details{
Generates output files by calling
\code{\link{MatchSigsAndRelabel}}
}
