% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignatureAnalyzer_interaction.R
\name{ReadAnySACat}
\alias{ReadAnySACat}
\title{Read a catalog of spectra or signatures in SignatureAnalyzer format.}
\usage{
ReadAnySACat(path, canonicalize.rows, exp.nrow, row.order)
}
\arguments{
\item{path}{File to read}

\item{canonicalize.rows}{A function to get the rownames and put the rows
in the correct order.}

\item{exp.nrow}{Expected number of rows. Stop if the actual
number of rows in path is different.}

\item{row.order}{Canonical row order for these mutation types.}
}
\value{
A catalog of spectra or signatures in canonical in-memory format.
}
\description{
Read a catalog of spectra or signatures in SignatureAnalyzer format.
}
\keyword{internal}
