% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignatureAnalyzerInteraction.R
\name{ReadAnySASig}
\alias{ReadAnySASig}
\title{Read signatures in SignatureAnalyzer format as found on
the PCAWG7 Synapse web site.}
\usage{
ReadAnySASig(path, canonicalize.rows, exp.nrow, row.order)
}
\arguments{
\item{path}{File to read}

\item{canonicalize.rows}{A function to get the rownames and put the rows
in the correct order.}

\item{exp.nrow}{Expected number of rows. Stop if the actual
number of rows in path is different.}

\item{row.order}{Canonical row order for these mutation types.}
}
\value{
A catalog of spectra or signatures in canonical in-memory format.
}
\description{
Read signatures in SignatureAnalyzer format as found on
the PCAWG7 Synapse web site.
}
\keyword{internal}
