% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiModalMuSigInteraction.R
\name{ReadCatMM}
\alias{ReadCatMM}
\title{Read Catalog files in MM format}
\usage{
ReadCatMM(cat)
}
\arguments{
\item{cat}{Input catalog, can be a tab-delimited text
file in MultiModalMuSig format.}
}
\value{
a catalog matrix in ICAMS format.
}
\description{
Read Catalog files in MM format
}
