% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadWriteExposures.R
\name{ReadExposure}
\alias{ReadExposure}
\title{Read an exposure matrix to a file}
\usage{
ReadExposure(file)
}
\arguments{
\item{file}{CSV file containing an exposure matrix}
}
\value{
Matrix of exposures
}
\description{
Read an exposure matrix to a file
}
