% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignatureAnalyzer_interaction.R
\name{ReadSACat}
\alias{ReadSACat}
\alias{ReadSACat96}
\alias{ReadSACat1536}
\alias{ReadSACatDBS}
\alias{ReadSACatID}
\title{Read Catalog Functions}
\usage{
ReadSACat96(path)

ReadSACat1536(path)

ReadSACatDBS(path)

ReadSACatID(path)
}
\arguments{
\item{path}{The file to read from.}
}
\value{
A catalog in canonical in-memory format.
}
\description{
Read a catalog or signature in SignatureAnalyzer format from path
}
\details{
\code{ReadSACat96} Read a 96 SNS catalog from path

\code{ReadSACat1536} Read a 1536 SNS catalog from path

\code{ReadSACatDBS} Read a 78 DNS catalog from path

\code{ReadSACatID} Read a ID (insertion/deletion) catalog from path
Please take note that deletion Repeat Size ranges from 0 to 5+
in the catalog, but for plotting and end user documentation
it ranges from 1 to 6+.
}
