% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignatureAnalyzerInteraction.R
\name{ReadSASig}
\alias{ReadSASig}
\alias{ReadSASig96}
\alias{ReadSASig1536}
\alias{ReadSASigDBS}
\alias{ReadSASigID}
\title{Read SignatureAnalyzer signatures from PCAWG7 / Synapse.}
\usage{
ReadSASig96(path)

ReadSASig1536(path)

ReadSASigDBS(path)

ReadSASigID(path)
}
\arguments{
\item{path}{The file to read from.}
}
\value{
A catalog in canonical in-memory format.
}
\description{
There is no custom catalog reading function for SignatureAnalyzer.
}
\details{
\code{ReadSASig96} Read a 96 SNS catalog from path

\code{ReadSASig1536} Read a 1536 SNS catalog from path

\code{ReadSASigDBS} Read a 78 DNS catalog from path

\code{ReadSASigID} Read a ID (insertion/deletion) catalog from path
Please take note that deletion Repeat Size ranges from 0 to 5+
in the catalog, but for plotting and end user documentation
it ranges from 1 to 6+.
}
\seealso{
\code{\link{ReadSASigCOMPOSITE}}
}
