% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignatureAnalyzerInteraction.R
\name{ReadSASigCOMPOSITE}
\alias{ReadSASigCOMPOSITE}
\title{Read SignatureAnalyzer COMPOSITE signatures from fixed files}
\usage{
ReadSASigCOMPOSITE()
}
\value{
A matrix (data.frame?) of COMPOSITE signatures
}
\description{
Read SignatureAnalyzer COMPOSITE signatures from fixed files
}
