% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignatureAnalyzerInteraction.R
\name{ReadSASigsCOMPOSITE}
\alias{ReadSASigsCOMPOSITE}
\title{Read SignatureAnalyzer COMPOSITE signatures from fixed files}
\usage{
ReadSASigsCOMPOSITE()
}
\value{
A matrix (data.frame?) of COMPOSITE signatures
}
\description{
Read SignatureAnalyzer COMPOSITE signatures from fixed files
}
