% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadWriteExposures.R
\name{ReadSynapseExposure}
\alias{ReadSynapseExposure}
\title{Read an exposure matrix from a Synapse file}
\usage{
ReadSynapseExposure(file)
}
\arguments{
\item{file}{CSV file containing an exposure matrix}
}
\value{
Matrix of exposures
}
\description{
Read an exposure matrix from a Synapse file
}
