% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RealExposures}
\alias{RealExposures}
\alias{sa.all.real.exposures}
\alias{sp.all.real.exposures}
\alias{sa.no.hyper.real.exposures}
\alias{sp.no.hyper.real.exposures}
\title{Real exposure (signature attributions) from SignatureAnalyzer and SigProfiler}
\format{Numerical matrix with rows indicating signatures and columns indicating
(tumor) samples.}
\source{
\url{https://dx.doi.org/10.7303/syn11761237.4}

\url{https://dx.doi.org/10.7303/syn11738669.5}

\url{https://dx.doi.org/10.7303/syn11761198.4}

\url{https://dx.doi.org/10.7303/syn11761237.4}
}
\usage{
sa.all.real.exposures

sp.all.real.exposures

sa.no.hyper.real.exposures

sp.no.hyper.real.exposures
}
\description{
Real exposure (signature attributions) from SignatureAnalyzer and SigProfiler
}
\note{
Prefix \code{sa} indicates SignatureAnalyzers, \code{sp} indicates
SigProfiler; \code{all} indicates all samples, \code{no.hyper} means
that hypermutated tumors as defined for SignatureAnalyzer have been removed.
}
\keyword{datasets}
