% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Runhdp.R
\name{Runhdp}
\alias{Runhdp}
\title{Run hdp extraction and attribution on a spectra catalog file}
\usage{
Runhdp(input.catalog, read.catalog.function, write.catalog.function,
  out.dir, CPU.cores = NULL, seedNumber = 1, K = NULL,
  K.range = NULL, test.only = FALSE, overwrite = FALSE)
}
\arguments{
\item{input.catalog}{File containing input spectra catalog.
Columns are samples (tumors), rows are mutation types.}

\item{read.catalog.function}{Function to read a catalog
(can be spectra or signature catalog): it takes a file path as
its only argument and returning a catalog as a numeric matrix.}

\item{write.catalog.function}{Function to write a catalog.}

\item{out.dir}{Directory that will be created for the output;
abort if it already exits.  Log files will be in
\code{paste0(out.dir, "/tmp")}.}

\item{CPU.cores}{Number of CPUs to use in running
hdp. For a server, 30 cores would be a good
choice; while for a PC, you may only choose 2-4 cores.
By default (CPU.cores = NULL), the CPU.cores would be equal
to \code{(parallel::detectCores())/2}, total number of CPUs
divided by 2.}

\item{seedNumber}{Specify the pseudo-random seed number
used to run hdp. Setting seed can make the
attribution of hdp repeatable.
Default: 1.}

\item{K, K.range}{\code{K} is the precise value for
the number of signatures active in spectra (K).
Specify \code{K} if you know precisely how many signatures
are active in the \code{input.catalog}, which is the
\code{ICAMS}-formatted spectra file.

\code{K.range} is A numeric vector \code{(K.min,K.max)}
of length 2 which tell hdp to search the best
signature number active in spectra, K, in this range of Ks.
Specify \code{K.range} if you don't know how many signatures
are active in the \code{input.catalog}.

WARNING: You must specify only one of \code{K} or \code{K.range}!

Default: NULL}

\item{test.only}{If TRUE, only analyze the first 10 columns
read in from \code{input.catalog}.
Default: FALSE}

\item{overwrite}{If TRUE, overwrite existing output.
Default: FALSE}
}
\value{
The attributed exposure of \code{hdp}, invisibly.
}
\description{
WARNING: hdp can only do exposure attribution
using SBS96 spectra catalog and signature catalog!
}
\details{
Creates several
 files in \code{out.dir}. These are:
 TODO(Steve): list the files

 TODO(Wuyang)
}
