% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignatureAnalyzerInteraction.R
\name{RunSignatureAnalyzerAttributeOnly}
\alias{RunSignatureAnalyzerAttributeOnly}
\title{Run SignatureAnalyzer attribution on a sa.output.rdata and its corresponding catalog file.}
\usage{
RunSignatureAnalyzerAttributeOnly(input.catalog, read.catalog.function,
  input.rdata, write.signature.function, out.dir, maxK = 30,
  tol = 1e-07, test.only = FALSE, input.exposures = NULL,
  delete.tmp.files = TRUE, overwrite = FALSE)
}
\arguments{
\item{input.catalog}{File containing input catalog.  Columns are
samples (tumors), rows are signatures.  SignatureAnalyzer does
not care about the row names (I think) TODO(Steve): check this.}

\item{read.catalog.function}{Function taking a file path as
its only argument and returning a catalog as a numeric matrix.}

\item{input.rdata}{Rdata file containing a list "out.data"
with 6 entries and a matrix "sigs", which is a matrix for
unnormalized signatures.}

\item{write.signature.function}{Function with first argument the
signatures generated by SignatureAnalyzer and second argument
the file to write to.}

\item{out.dir}{Directory that will be created for the output;
abort if it already exits.  Log files will be in
\code{paste0(out.dir, "/tmp")}.}

\item{maxK}{The maximum number of signatures to consider
extracting.}

\item{tol}{Controls when SignatureAnalyzer will terminate
its search; \code{tol} was 1.e-05 for the PCAWG7 analysis.}

\item{test.only}{If TRUE, only analyze the first 10 columns
read in from \code{input.catalog}.}

\item{input.exposures}{A file with the synthetic exposures used to generate
\code{input.catalog}; if provided here,
this is copied over to the output directory
for downstream analysis.}

\item{delete.tmp.files}{If TRUE delete the many temporary
files generated by SignatureAnalyzer.}

\item{overwrite}{If TRUE, overwrite existing output}
}
\value{
The output of SignatureAnalyzer, invisibly. This is a list
with five elements (named in this code, not by SignatureAnalyzer):
\enumerate{
 \item \code{signatures.W} The signature matrix
 \item \code{exposures.fine.tuned} The corresponding exposures
 \item \code{likelihood} The likelihood
 \item \code{evidence} -1 * the posterior probability
 \item \code{relevance} One for each column of the \code{signatures.W}
 \item \code{error} A measure of reconstruction error (?)
}
}
\description{
Normally, please call \code{\link{SignatureAnalyzerOneRun}}
instead of this function.
}
\details{
Creates several
 files in \code{paste0(out.dir, "/sa.output.rdata")}. These are
 TODO(Steve): list the files
}
