% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunSignatureAnalyzerAttribution.R
\name{RunSignatureAnalyzerAttribution}
\alias{RunSignatureAnalyzerAttribution}
\title{Run SignatureAnalyzer attribution on a catalog file and output by RunSignatureAnalyzerOnFile().}
\usage{
RunSignatureAnalyzerAttribution(input.catalog, read.catalog.function,
  extracted.signature.file, raw.exposures.file, write.signature.function,
  out.dir, test.only = FALSE, input.exposures = NULL,
  delete.tmp.files = TRUE, overwrite = FALSE)
}
\arguments{
\item{input.catalog}{File containing input catalog.  Columns are
samples (tumors), rows are signatures.  SignatureAnalyzer does
not care about the row names (I think) TODO(Steve): check this.}

\item{read.catalog.function}{Function taking a file path as
its only argument and returning a catalog as a numeric matrix.}

\item{extracted.signature.file}{A .csv file containing extracted
signatures. Normally, this file is named \code{"sa.output.sigs.csv"}
and is generated by function \code{RunSignatureAnalyzerOnFile()}.
It expects to have the same format as the \code{input.catalog},
thus it will be read by \code{read.catalog.function} too.}

\item{raw.exposures.file}{A .csv file containing raw attributions
of exposures. Normally, this file is named \code{"sa.output.raw.exp.csv"}
and is generated by function \code{RunSignatureAnalyzerOnFile()}.}

\item{write.signature.function}{Function with first argument the
signatures generated by SignatureAnalyzer and second argument
the file to write to.}

\item{out.dir}{Directory that will be created for the output;
abort if it already exits.  Log files will be in
\code{paste0(out.dir, "/tmp")}.}

\item{test.only}{If TRUE, only analyze the first 10 columns
read in from \code{input.catalog}.}

\item{input.exposures}{A file with the synthetic exposures used to generate
\code{input.catalog}; if provided here,
this is copied over to the output directory
for downstream analysis.}

\item{delete.tmp.files}{If TRUE delete the many temporary
files generated by SignatureAnalyzer.}

\item{overwrite}{If TRUE, overwrite existing output}
}
\value{
The final attribution matrix.
(i.e. \code{exp.fine.tuned})
}
\description{
Normally, please call \code{\link{SignatureAnalyzerOneRun}}
instead of this function.
}
\details{
Save the final attribution of a catalog matrix into
a file named \code{"sa.output.fine.exp.csv"} under the
folder \code{out.dir}.
}
