% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move-contents-to-signatureanalyzerinteraction.R
\name{RunSignatureAnalyzerOnFile}
\alias{RunSignatureAnalyzerOnFile}
\title{for COMPOSITE signatures we rbind coordinated
"SNV", "DNP", and "INDEL" signatures. See
\code{\link{ReadSASigsCOMPOSITE}}.
Function for running SignatureAnalzyer on a file containing
a catalog.}
\usage{
RunSignatureAnalyzerOnFile(input.catalog, read.catalog.function, out.dir,
  write.signature.function, maxK = 30, tol = 1e-07,
  test.only = FALSE)
}
\arguments{
\item{input.catalog}{File containing input catalog.  Columns are
samples (tumors), rows are signatures.  SignatuerAnalyzer does
not care about the row names (I think) TODO(Steve): check this.}

\item{read.catalog.function}{Function taking a file path as
its only argument and returning a catalog as a numeric matrix.}

\item{out.dir}{Directory that will be created for the output;
abort if it already exits.  Log files will be in
\code{paste0(out.dir, "/tmp")}.}

\item{write.signature.function}{Function with first argument the
signatures generated by SignatureAnalyzer and second argument
the file to write to.}

\item{maxK}{The maximun number of signatures to consider
extracting.}

\item{tol}{Controls when SignatureAnalyzer will terminate
its search; \code{tol} was 1.e-05 for the PCAWG7 analysis.}

\item{test.only}{If true, only analyze the first 10 columns
read in from \code{input.catalog}.}
}
\value{
The output of SignatureAnalyzer, invisibily.
}
\description{
It is necessary to setwd() to a folder containing
SignatureAnalyzer (right now SignatureAnalzyer.052418)
and then call
\preformatted{
options( warn = 0 )
here <- getwd()
setwd("SignatureAnalzyer.052418/") # Or the appropriate directory
INPUT <- "INPUT_SignatureAnalyzer/"
source("SignatureAnalyzer.PCAWG.function.R")
setwd(here) # This is necessary because the caller
            # as specified input and output locations
            # realtive to here.
RunSignatureAnalyzerOnSyn(...)
}
}
\details{
TODO(Steve): see if we can do the source of SA inside
this function.

Creates the file \code{paste0(out.dir, "/sa.output.rdata")}.

TODO(Steve): review the output format from SignatureAnalyzer -- names of sigs just come from the input?
}
