% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunsigneR.R
\name{RunsigneR}
\alias{RunsigneR}
\title{Run signeR extraction and attribution on a spectra catalog file}
\usage{
RunsigneR(input.catalog, read.catalog.function, write.catalog.function,
  out.dir, seedNumber = 1, K = NULL, K.range = NULL,
  test.only = FALSE, overwrite = FALSE)
}
\arguments{
\item{input.catalog}{File containing input spectra catalog.
Columns are samples (tumors), rows are mutation types.}

\item{read.catalog.function}{Function to read a catalog
(can be spectra or signature catalog): it takes a file path as
its only argument and returning a catalog as a numeric matrix.}

\item{write.catalog.function}{Function to write a catalog.}

\item{out.dir}{Directory that will be created for the output;
abort if it already exits.  Log files will be in
\code{paste0(out.dir, "/tmp")}.}

\item{seedNumber}{Specify the pseudo-random seed number
used to run signeR. Setting seed can make the
attribution of signeR repeatable.
Default: 1.}

\item{K, K.range}{\code{K} is the precise value for
the number of signatures active in spectra (K).
Specify \code{K} if you know precisely how many signatures
are active in the \code{input.catalog}, which is the
\code{ICAMS}-formatted spectra file.

\code{K.range} is A numeric vector \code{(K.min,K.max)}
of length 2 which tell signeR to search the best
signature number active in spectra, K, in this range of Ks.
Specify \code{K.range} if you don't know how many signatures
are active in the \code{input.catalog}.

WARNING: You must specify only one of \code{K} or \code{K.range}!

Default: NULL}

\item{test.only}{If TRUE, only analyze the first 10 columns
read in from \code{input.catalog}.
Default: FALSE}

\item{overwrite}{If TRUE, overwrite existing output.
Default: FALSE}
}
\value{
The attributed exposure of \code{signeR}, invisibly.
}
\description{
Run signeR extraction and attribution on a spectra catalog file
}
\details{
Creates several
 files in \code{out.dir}. These are:
 TODO(Steve): list the files

 TODO(Wuyang)
}
