% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunYAPSAAttributionOnly.R
\name{RunYAPSAAttributeOnly}
\alias{RunYAPSAAttributeOnly}
\title{Run YAPSA attribution on a spectra catalog file.}
\usage{
RunYAPSAAttributeOnly(input.catalog, gt.sigs.file, read.catalog.function,
  out.dir, seed = 1, signature.cutoff = NULL, test.only = FALSE,
  input.exposures = NULL, delete.tmp.files = TRUE, overwrite = FALSE)
}
\arguments{
\item{input.catalog}{File containing input spectra catalog.
Columns are samples (tumors), rows are mutation types.}

\item{gt.sigs.file}{File containing input mutational signatures.
Columns are signatures, rows are mutation types.}

\item{read.catalog.function}{Function taking a file path as
its only argument and returning a catalog as a numeric matrix.}

\item{out.dir}{Directory that will be created for the output;
abort if it already exits.  Log files will be in
\code{paste0(out.dir, "/tmp")}.}

\item{seed}{Specify the pseudo-random seed number
used to run YAPSA. Setting seed can make the
attribution of YAPSA repeatable.
Default: 1.}

\item{signature.cutoff}{A numeric vector of values less than 1.
Signatures from within W with an overall exposure
less than the respective value in \code{in_cutoff_vector}
will be discarded.
Default: vector length of number of sigs with all zeros}

\item{test.only}{If TRUE, only analyze the first 10 columns
read in from \code{input.catalog}.
Default: FALSE}

\item{input.exposures}{A file with the synthetic exposures
used to generate \code{input.catalog}; if provided here,
this is copied over to the output directory
for downstream analysis.}

\item{overwrite}{If TRUE, overwrite existing output.
Default: FALSE}
}
\value{
The attributed exposure of \code{deconstructSigs}, invisibly.
}
\description{
Run YAPSA attribution on a spectra catalog file.
}
\details{
Creates several
 files in \code{paste0(out.dir, "/sa.output.rdata")}. These are
 TODO(Steve): list the files
}
