% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignatureAnalyzer_interaction.R
\name{SACat96}
\alias{SACat96}
\title{Prepare input data for SignatureAnalyzer}
\usage{
SACat96(cat96)
}
\arguments{
\item{cat96}{Input catalog in in standard in-memory format}
}
\value{
Catalog in in-memory format expected by SignatureAnalyzer
}
\description{
Prepare input data for SignatureAnalyzer
}
