% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignatureAnalyzerInteraction.R
\name{SAMultiRunOneCatalog}
\alias{SAMultiRunOneCatalog}
\title{Run SignatureAnalyzer many times on one catalog and put results
in specified location.}
\usage{
SAMultiRunOneCatalog(num.runs, signatureanalyzer.code.dir, input.catalog,
  read.catalog.function, out.dir, write.signature.function, maxK = 30,
  tol = 1e-07, test.only = FALSE, delete.tmp.files = TRUE,
  overwrite = FALSE, mc.cores = 1, verbose = FALSE, seed = NULL)
}
\arguments{
\item{num.runs}{The number of times run SignatureAnalyzer on each
catalog (matrix of mutational spectra).}

\item{signatureanalyzer.code.dir}{The directory holding the
SignatureAnalyzer code.}

\item{input.catalog}{The catalog to analyze.}

\item{read.catalog.function}{Function taking a file path as
its only argument and returning a catalog as a numeric matrix.}

\item{out.dir}{Root of directory tree that will contain the
results.}

\item{write.signature.function}{Function with first argument the
signatures generated by SignatureAnalyzer and second argument
the file to write to.}

\item{maxK}{The maximum number of signatures to consider
extracting.}

\item{tol}{Controls when SignatureAnalyzer will terminate
its search; \code{tol} was 1.e-05 for the PCAWG7 analysis.}

\item{test.only}{If TRUE, only analyze the first 10 columns
read in from \code{input.catalog}.}

\item{delete.tmp.files}{If TRUE delete the many temporary
files generated by SignatureAnalyzer.}

\item{overwrite}{If TRUE overwrite previous results in same directory tree.}

\item{mc.cores}{Number of cores to use for \code{mclapply}; ignored on
Windows.}

\item{verbose}{If TRUE cat a message regarding progress.}

\item{seed}{If not \code{NULL} call
\code{RNGkind(kind = "L'Ecuyer-CMRG"); set.seed(seed)}.}
}
\description{
Run SignatureAnalyzer many times on one catalog and put results
in specified location.
}
