% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateSynData.R
\name{SetNewOutDir}
\alias{SetNewOutDir}
\title{Globally set the location used by \code{OutDir}.}
\usage{
SetNewOutDir(dir)
}
\arguments{
\item{dir}{The location}
}
\description{
Globally set the location used by \code{OutDir}.
}
