% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SigMisc.R
\docType{package}
\name{SigMisc}
\alias{SigMisc}
\alias{SigMisc-package}
\title{SigMisc: A package for working with mutational signatures}
\description{
The SigMisc package provides several categories
of functions for working with mutational signatures.
}
\section{Signature Matching and Comparison}{

Functions for comparing mutational signatures and
sets of mutational signatures:
\code{\link{Match1Sig}},
\code{\link{MatchSigs1Direction}},
\code{\link{MatchSigs2Directions}},
\code{\link{MatchSigsThenWriteAndPlot}},
\code{\link{TestMatchSigs2Directions}}
}

\section{Miscellaneous}{

fwriteDataFrame
}

\keyword{internal}
