% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignatureAnalyzerInteraction.R
\name{SignatureAnalyzer4MatchedCatalogs}
\alias{SignatureAnalyzer4MatchedCatalogs}
\title{Run SignatureAnalyzer on 4 coordinated data sets and put results
in specified location.}
\usage{
SignatureAnalyzer4MatchedCatalogs(num.runs = 20,
  signatureanalyzer.code.dir, dir.root, maxK = 30, tol = 1e-07,
  test.only = FALSE, delete.tmp.files = TRUE, slice = 1:4,
  overwrite = FALSE, mc.cores = 1)
}
\arguments{
\item{num.runs}{Number of SignatureAnalyzer runs per data set.}

\item{signatureanalyzer.code.dir}{The directory holding the
SignatureAnalyzer code.}

\item{dir.root}{Root of directory tree that contains the
input data and to which the results will be written.}

\item{maxK}{The maximum number of signatures to consider
extracting.}

\item{tol}{Controls when SignatureAnalyzer will terminate
its search; \code{tol} was 1.e-05 for the PCAWG7 analysis.}

\item{test.only}{If TRUE, only analyze the first 10 columns
read in from \code{input.catalog}.}

\item{delete.tmp.files}{If TRUE delete the many temporary
files generated by SignatureAnalyzer.}

\item{slice}{Vector of integers from 1:4. Only run on the
corresponding data set (see Details).}

\item{overwrite}{if TRUE overwrite preexisting results.}

\item{mc.cores}{The number of cores to use with \code{mclapply};
automatically overridden to 1 on Windows.}
}
\description{
Run SignatureAnalyzer on 4 coordinated data sets and put results
in specified location.
}
\details{
The 4 coordinated data sets are

\enumerate{
\item \code{sa.sa.96}

\item \code{sp.sp}

\item \code{sa.sa.COMPOSITE}

\item \code{sp.sa.COMPOSITE}

}
which are described elsewhere.
}
