% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignatureAnalyzerInteraction.R
\name{SignatureAnalyzerOneRun}
\alias{SignatureAnalyzerOneRun}
\title{Run SignatureAnalyzer once on a single data set and put results
in specified location.}
\usage{
SignatureAnalyzerOneRun(signatureanalyzer.code.dir, input.catalog,
  read.catalog.function, out.dir, write.signature.function,
  input.exposures = NULL, maxK = 30, tol = 1e-07,
  test.only = FALSE, delete.tmp.files = TRUE, verbose = 1,
  overwrite = FALSE)
}
\arguments{
\item{signatureanalyzer.code.dir}{The directory holding the
SignatureAnalyzer code.}

\item{input.catalog}{Path to the file containing the catalog
of mutational spectra.}

\item{read.catalog.function}{Function to read \code{input.catalog}.}

\item{out.dir}{Location to put the output; must not already exist; it
will be created.}

\item{write.signature.function}{Function to write the extracted
signatures to a file.}

\item{input.exposures}{Exposures from which the spectra in
\code{input.catalog} were generated.}

\item{maxK}{The maximum number of signatures to consider
extracting.}

\item{tol}{Controls when SignatureAnalyzer will terminate
its search; \code{tol} was 1.e-05 for the PCAWG7 analysis.}

\item{test.only}{If TRUE, only analyze the first 10 columns
read in from \code{input.catalog}.}

\item{delete.tmp.files}{If TRUE delete the many temporary
files generated by SignatureAnalyzer.}

\item{verbose}{If > 0, write some tracing and timing information.}

\item{overwrite}{If TRUE overwrite preexisting results.

#' @return The output of SignatureAnalyzer, invisibly. This is a list
with five elements (named in this code, not by SignatureAnalyzer):
\enumerate{
 \item \code{signatures.W} The signature matrix
 \item \code{exposures.H} The corresponding exposures
 \item \code{likelihood} The likelihood
 \item \code{evidence} -1 * the posterior probability
 \item \code{relevance} One for each column of the \code{signatures.W}
 \item \code{error} A measure of reconstruction error (?)
}}
}
\description{
Run SignatureAnalyzer once on a single data set and put results
in specified location.
}
