% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignatureAnalyzerHyperRun.R
\name{SignatureAnalyzerPrepHyper1Secondary}
\alias{SignatureAnalyzerPrepHyper1Secondary}
\title{Prepare the "hypermutated" segment (a.k.a "Secondary" segment
of a split non-hyper and hyper data set.)}
\usage{
SignatureAnalyzerPrepHyper1Secondary(non.hyper.results, primary.catalog,
  hyper.catalog, secondary.catalog, read.fn, write.fn)
}
\arguments{
\item{non.hyper.results}{The directory containing the the results of the analysis
of the non-hyper-mutated (a.k.a "PRIMARY") mutational spectra.}

\item{primary.catalog}{The catalog of non-hyper-mutated
mutational spectra from which the results in \code{non.hyper.results}
were derived.}

\item{hyper.catalog}{The catalog of hyper-mutated
mutational spectra which will be part of the input for the
secondary analysis.}

\item{secondary.catalog}{The final output catalog on which
the secondary analysis will be performed; this is a
\code{cbind} of pseudo-spectra generated from the PRIMARY
signatures with the \code{hyper.catalog}.}

\item{read.fn}{Function to use for reading signatures.}

\item{write.fn}{Function to use for writing signatures.}
}
\description{
Prepare the "hypermutated" segment (a.k.a "Secondary" segment
of a split non-hyper and hyper data set.)
}
