% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignatureAnalyzerHyperRun.R
\name{SignatureAnalyzerPrepHyper4}
\alias{SignatureAnalyzerPrepHyper4}
\title{Prepare the "hypermutated" segment (a.k.a "Secondary" segment
of a split non-hyper and hyper data set.)}
\usage{
SignatureAnalyzerPrepHyper4(parent.dir, overwrite = FALSE)
}
\arguments{
\item{parent.dir}{A directory that must contain subdirectories
\code{syn.SA.hyper.low} and \code{syn.SA.hyper.mixed}.
\code{syn.SA.hyper.low} must contain the synthetic
non-hypermutated data and the results of running SignatureAnalyzer on the non-hyper
segment, with subdirectories \code{sa.sa.96}, \code{sa.sa.COMPOSITE},
\code{sp.sa.COMPOSITE}, and \code{sp.sp}.
\code{syn.SA.hyper.mixed} must contain the synthetic hypermutated data. The results of
the initial SignatureAnalyzer run will be placed here to prepare this directory
for the second SignatureAnalyzer run.}

\item{overwrite}{If \code{TRUE} overwrite existing directories and files.}
}
\description{
Prepare the "hypermutated" segment (a.k.a "Secondary" segment
of a split non-hyper and hyper data set.)
}
