% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{SignatureAnalyzerSigs}
\alias{SignatureAnalyzerSigs}
\title{Signatures from SignatureAnalyzer}
\format{Numerical matrix with rows indicating mutation
types and columns indicating signatures.}
\description{
Signatures from SignatureAnalyzer
}
