% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummarizeResultsSA.R
\name{SignatureAnalyzerSummarizeTopLevel}
\alias{SignatureAnalyzerSummarizeTopLevel}
\title{Summarize all subdirectories of SignatureAnalyzer results
on a major dataset.}
\usage{
SignatureAnalyzerSummarizeTopLevel(top.level.dir, overwrite = FALSE)
}
\arguments{
\item{top.level.dir}{Path to top level directory, which
must contain the following subdirectories:
\itemize{
\item \code{sa.sa.96/sa.results/}
\item \code{sp.sp/sa.results/}
\item \code{sa.sa.COMPOSITE/sa.results/}
\item \code{sp.sa.COMPOSITE/sa.results/}
}
Each of the directories must contain
additonal subdirectories, one for each SignatureAnalyzer
run, names \code{sa.run.<n>}, where <n> is an integer
(string of digits).}

\item{overwrite}{If TRUE overwrite existing summary files.}
}
\description{
This function depends on a particular directory structure: see
argument \code{top.level.dir}. This function finds the
best of multiple SignatureAnalyzer extraction runs and summarizes
the comparision of the best run with the ground truth.
}
