% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SigSimilarity.R
\name{SigSetSimilarity}
\alias{SigSetSimilarity}
\title{Calculate bidirectional closest similarities bewteen two sets of signatures
and the average of the similarities.}
\usage{
SigSetSimilarity(sigs1, sigs2)
}
\arguments{
\item{sigs1}{Matrix of signatures; colnames used as identifiers}

\item{sigs2}{Matrix of signatures; colnames used as identifiers}
}
\value{
#  A list with the elements:
   \code{avg}: the average of the cosine similarities between each signature
        in sigs1 and its closest match in sigs2 and the closest match
        between each signature in sigs2 and its closest match in sigs1

   \code{match1}: a data frame with a signature in sigs1 in the first column,
           the closest match in sigs2 in the second column, and the
           cosine similarity between them in the third column; the rownames
           are the values in column 1.

   \code{match2}: a data frame with a signature in sigs2 in the first column,
           the closest match in sigs1 in the second column, and the
           cosine similarity between them in the third column; the rownames
           are the values in column 1.

\code{match1} and \code{match2} might not have the same number of rows.
}
\description{
Calculate bidirectional closest similarities bewteen two sets of signatures
and the average of the similarities.
}
