% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignatureAnalyzerInteraction.R
\name{SourceSignatureAnlyzerCode}
\alias{SourceSignatureAnlyzerCode}
\title{Source SignatureAnalyzer Codes.}
\usage{
SourceSignatureAnlyzerCode(signatureanalyzer.code.dir)
}
\arguments{
\item{signatureanalyzer.code.dir}{The directory which stores
SignatureAnalyzer program files. It must include a folder
named \code{INPUT_SignatureAnalyzer} and a R script named
\code{SignatureAnalyer.PCAWG.function.R}}
}
\description{
Source SignatureAnalyzer Codes.
}
