% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadWriteCatalogs.R
\name{SplitCatCOMPOSITE}
\alias{SplitCatCOMPOSITE}
\title{Split COMPOSITE (SNS1536+DBS78+ID83) catalogues
in ICAMS format into 3 individual catalogs.}
\usage{
SplitCatCOMPOSITE(catalog)
}
\arguments{
\item{catalog}{Input catalog, can be a .csv file or matrix
in ICAMS COMPOSITE format.}
}
\value{
a list, containing 3 catalog matrices in MultiModalMuSig format.
Each matrix contains SNS1536, DBS78 and ID83 information, respectively.
}
\description{
Split COMPOSITE (SNS1536+DBS78+ID83) catalogues
in ICAMS format into 3 individual catalogs.
}
