% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummarizeResultsOtherPackages.R
\name{SummarizeSigOneAttr96Subdir}
\alias{SummarizeSigOneAttr96Subdir}
\title{Assess/evaluate SBS96 results from packages which can
ONLY do exposure attribution.}
\usage{
SummarizeSigOneAttr96Subdir(third.level.dir,
  ground.truth.exposure.name = "ground.truth.syn.exposures.csv",
  overwrite = FALSE)
}
\arguments{
\item{third.level.dir}{Lowest level path to results, e.g.
\code{<top.dir>}\code{/sa.sa.96/SomaticSignatures.results/}
Here, \code{<top.dir>} refers to a top-level directory which contains the
full information of a synthetic dataset. (e.g. \code{syn.2.7a.7b.abst.v8})
This code depends on a conventional directory structure documented
elsewhere. For packages which can do both extraction and attribution,
we expect two files, \code{ground.truth.signatures.csv}
and \code{attributed.exposures.csv} are in the folder.}

\item{ground.truth.exposure.name}{File name which stores ground-truth exposures;
defaults to \code{"ground.truth.syn.exposures.csv"}.
This file can be found in the \code{sub.dir}, i.e. \code{<third.level.dir>/../}}

\item{overwrite}{If TRUE overwrite existing directories and files.}
}
\description{
Packages including but not limited to:
deconstructSigs, YAPSA.
}
\details{
Here, we excluded SignatureEstimation. Although it is also
a package with only attribution, but it has two attribution
algorithms. Therefore the naming of the results are slightly
different from the other two packages.
}
