% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummarizeResults.R
\name{SummarizeSigOneSA96Subdir}
\alias{SummarizeSigOneSA96Subdir}
\title{Summarize 96-channel results from SignatureAnalyzer}
\usage{
SummarizeSigOneSA96Subdir(third.level.dir,
  ground.truth.exposure.name = "ground.truth.syn.exposures.csv",
  which.run = "/best.run/", overwrite = FALSE)
}
\arguments{
\item{third.level.dir}{Lowest level path to results, that is
\code{<top.dir>}\code{/sa.sa.96/sa.results/},
\code{<top.dir>}\code{/sp.sp/sa.results/},
\code{<top.dir>}\code{/sa.sa.COMPOSITE/sa.results/}, or
\code{<top.dir>}\code{/sp.sa.COMPOSITE/sa.results/}.
Here, \code{<top.dir>} refers to a top-level directory which contains the
full information of a synthetic dataset. (e.g. \code{syn.2.7a.7b.abst.v8})
This code depends on a conventional directory structure documented
elsewhere. However there should be a directory
\code{<third.level.dir>}\code{/SBS96} which
stores SigProfiler results.}

\item{ground.truth.exposure.name}{File name which stores ground-truth exposures;
defaults to \code{"ground.truth.syn.exposures.csv"}.
This file can be found in the \code{sub.dir}, i.e. \code{<third.level.dir>/../}}

\item{which.run}{Name of subdirectory containing the run to summarize.}

\item{overwrite}{If TRUE overwrite existing directories and files.}
}
\description{
Summarize 96-channel results from SignatureAnalyzer
}
