% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummarizeResultsSE.R
\name{SummarizeSigOneSE96Subdir}
\alias{SummarizeSigOneSE96Subdir}
\title{Summarize 96-channel results from SignatureEstimation.}
\usage{
SummarizeSigOneSE96Subdir(third.level.dir,
  ground.truth.exposure.name = "ground.truth.syn.exposures.csv",
  overwrite = FALSE)
}
\arguments{
\item{third.level.dir}{Lowest level path to results, e.g.
\code{<top.dir>}\code{/sa.sa.96/SignatureEstimation.results/} or
\code{<top.dir>}\code{/sp.sp/SignatureEstimation.results/}
Here, \code{<top.dir>} refers to a top-level directory which contains the
full information of a synthetic dataset. (e.g. \code{syn.2.7a.7b.abst.v8})
This code depends on a conventional directory structure documented
elsewhere. However in \code{<third.level.dir>} there should be files
\code{attributed.exposures.QP.csv} and \code{attributed.exposures.SA.csv}
which store SignatureEstimation results.}

\item{ground.truth.exposure.name}{File name which stores ground-truth exposures;
defaults to \code{"ground.truth.syn.exposures.csv"}.
This file can be found in the \code{sub.dir}, i.e. \code{<third.level.dir>/../}}

\item{overwrite}{If TRUE overwrite existing directories and files.}
}
\description{
Summarize 96-channel results from SignatureEstimation.
}
