% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SynSig.R
\docType{package}
\name{SynSig}
\alias{SynSig}
\alias{SynSig-package}
\title{SynSig: A package for working with synthetic mutational signatures for
purpose of evalating software packages for extracting and attributing
mutational signatures.}
\description{
The SynSig package provides several categories
of functions for working with mutational signatures.
}
\section{Creating Synthetic Mutational Catalogs}{


The functions create synthetic mutational catalogs based
on parameters derived from signature profiles
and attributions (exposures)
}

\section{Signature Matching and Comparison}{


Functions for comparing mutational signatures and
sets of mutational signatures:
\code{\link{Match1Sig}},
\code{\link{MatchSigs1Direction}},
\code{\link{MatchSigs2Directions}},
\code{\link{MatchSigsThenWriteAndPlot}},
\code{\link{TestMatchSigs2Directions}}
}

\section{Miscellaneous}{

fwriteDataFrame
}

\keyword{internal}
