% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SynSigFunctions.v0.7.R
\name{SynSigParamsOneSignature}
\alias{SynSigParamsOneSignature}
\title{Extract SynSig parameters for one mutational signature profile}
\usage{
SynSigParamsOneSignature(counts, target.size)
}
\arguments{
\item{counts}{A vector of mutation counts attributed to one signature across
length(counts) samples.}

\item{target.size}{The length of genomic sequence from which the counts
were dervived, in megabases TODO:(steve) probably do not need this}
}
\value{
A 3-element vector with names "prevalence", "mean", and "stdv"
}
\description{
Extract SynSig parameters for one mutational signature profile
}
