% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SigSimilarity.R
\name{TestMatchSigs2Directions}
\alias{TestMatchSigs2Directions}
\title{Test MatchSigs2Directions}
\usage{
TestMatchSigs2Directions()
}
\description{
Test MatchSigs2Directions
}
\details{
Stop on error
}
\seealso{
Other signature matching functions: \code{\link{Match1Sig}},
  \code{\link{MatchSigs1Direction}},
  \code{\link{MatchSigs2Directions}},
  \code{\link{MatchSigsThenWriteAndPlot}}
}
\concept{signature matching functions}
