% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignatureAnalyzer_interaction.R
\name{WriteCatCOMPOSITE}
\alias{WriteCatCOMPOSITE}
\title{Write a SignatureAnalyzer catalog or signature matrix to disk}
\usage{
WriteCatCOMPOSITE(ct, path)
}
\arguments{
\item{ct}{A catalog or signature matrix}

\item{path}{Path to file to write}

\item{rowname.name}{Name of first column in file}
}
\description{
Write a SignatureAnalyzer catalog or signature matrix to disk
}
