% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadWriteExposures.R
\name{WriteExposure}
\alias{WriteExposure}
\title{Write exposure matrix to a file}
\usage{
WriteExposure(exposure.matrix, file)
}
\arguments{
\item{exposure.matrix}{Matrix of exposures}

\item{file}{File to which to write the exposure matrix (as a CSV file)}
}
\description{
Write exposure matrix to a file
}
