% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateSynData.R
\name{AddAllScripts}
\alias{AddAllScripts}
\title{Create scripts to run SignatureAnalyzer in all subdirectories
of \code{OutDir()}.}
\usage{
AddAllScripts(maxK = 30)
}
\arguments{
\item{maxK}{The \code{maxK} argument for SignatureAnalyzer.}
}
\description{
Create scripts to run SignatureAnalyzer in all subdirectories
of \code{OutDir()}.
}
