% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateSynData.R
\name{AddAllScripts}
\alias{AddAllScripts}
\title{Use the current value of \code{OutDir()} to copy scripts into all subdirectories}
\usage{
AddAllScripts(maxK = 30)
}
\arguments{
\item{maxK}{The \code{maxK} argument for SignatureAnalyzers.}
}
\description{
Use the current value of \code{OutDir()} to copy scripts into all subdirectories
}
