% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BestSignatureAnalyzerResult.R
\name{CopyBestSignatureAnalyzerResult}
\alias{CopyBestSignatureAnalyzerResult}
\title{Find the SignatureAnalyzer results
directory with the best results and make a copy
of it as \code{sa.results.dir/best.run/}}
\usage{
CopyBestSignatureAnalyzerResult(sa.results.dir, verbose = FALSE,
  overwrite = FALSE)
}
\arguments{
\item{sa.results.dir}{See \code{\link{BestSignatureAnalyzerResult}}}

\item{verbose}{See \code{\link{BestSignatureAnalyzerResult}}}

\item{overwrite}{If TRUE overwrite existing "best.run"}
}
\value{
The path of the best directory that was copied as a
string, with the list directories examined as the attribute
\code{run.directories}.
}
\description{
Find the SignatureAnalyzer results
directory with the best results and make a copy
of it as \code{sa.results.dir/best.run/}
}
