% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateRandomSynData.R
\name{CreateExposuresNums}
\alias{CreateExposuresNums}
\title{Create a \code{num.exposures} random exposures
from a normal distribution with \code{mean} and
\code{sd}.}
\usage{
CreateExposuresNums(num.exposures, mean, sd, total.num.sigs)
}
\arguments{
\item{num.exposures}{Number of exposures to create.}

\item{mean}{Mean of distribution to draw from.}

\item{sd}{Stadard deviation of distribution to draw from.}

\item{total.num.sigs}{Numer of signatures in the "universe".}
}
\description{
Discard 0 exposures and exposures > than \code{max.exposures}.
}
