% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateTotalRandomSynData2.R
\name{CreateOneRandomMutSigProfile}
\alias{CreateOneRandomMutSigProfile}
\title{Create one "random" signature profile.}
\usage{
CreateOneRandomMutSigProfile(row.headers)
}
\arguments{
\item{row.headers}{One of the \code{\link{ICAMS}} package variable such as
\code{\link{.catalog.row.order96}}.}
}
\value{
A single column matrix with rownames \code{row.headers} and
  colname \code{RandSig}.
}
\description{
Create one "random" signature profile.
}
\keyword{internal}
