% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Create.syn.30.random.R
\name{CreateOneRandomMutSigProfile}
\alias{CreateOneRandomMutSigProfile}
\title{Create one "random" artificial signature profile.}
\usage{
CreateOneRandomMutSigProfile(row.names)
}
\arguments{
\item{row.names}{One of the \code{\link{ICAMS}} package variable such as
\code{catalog.row.order[["SBS96"]]}.}
}
\value{
A single column matrix with \code{rownames} \code{row.headers} and
  \code{colnames} \code{"RandSig"}.
}
\description{
Create one "random" artificial signature profile.
}
\keyword{internal}
