% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Create.syn.30.random.R
\name{CreateRandomSAAndSPSynCatalogs}
\alias{CreateRandomSAAndSPSynCatalogs}
\title{Create a full SignatureAnalyzer / SigProfiler test data set for "random"
artificial signatures.}
\usage{
CreateRandomSAAndSPSynCatalogs(top.level.dir, num.syn.tumors,
  overwrite = FALSE)
}
\arguments{
\item{top.level.dir}{Path to top level of directory structure to be created.}

\item{num.syn.tumors}{Number of synthetic tumors to create.}

\item{overwrite}{If TRUE, overwrite existing directories / files.}
}
\description{
Create a full SignatureAnalyzer / SigProfiler test data set for "random"
artificial signatures.
}
