% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Create.syn.SBS1.5.correlated.R
\name{CreateSBS1SBS5CorrelatedSyntheticData}
\alias{CreateSBS1SBS5CorrelatedSyntheticData}
\title{Wrapper function for generating SBS1-SBS5-correlated Synthetic data}
\usage{
CreateSBS1SBS5CorrelatedSyntheticData(dir.name = "./S.0.5.Rsq.0.3",
  dataset.name = NULL, overwrite = FALSE, seed = 1,
  main.signature = "SBS5", correlated.signature = "SBS1",
  name.prefix = "TwoCorreSigsGen", sample.number = 500,
  main.mean.log = 2.5, main.stdev.log = 0.3,
  correlated.stdev.log = 0.4, slope.linear = 0.5,
  main.signature.lower.thres = 100,
  correlated.signature.lower.thres = 1, pearson.r.2.lower.thres = 0.29,
  pearson.r.2.higher.thres = 0.31,
  min.main.to.correlated.ratio.linear = 1/3,
  max.main.to.correlated.ratio.linear = Inf)
}
\arguments{
\item{dir.name}{Folder to place the generated tumor spectra
and other output files.
Default: ./S.0.5.Rsq.0.3}

\item{dataset.name}{The dataset.name encodes the parameters for
the synthetic data, but this is just a convention.
If NULL, it will be changed to the last part of the \code{dir.name}
(Default: NULL)}

\item{overwrite}{Whether to overwrite
(Default: FALSE)}

\item{seed}{The seed number used to initizalize pesudo-random number
generator (RNG). This makes the generation of the correlated
datasets repeatable. (Default: 1)}

\item{main.signature}{The name of the main signature whose exposure
can vary freely. (Default: SBS5)}

\item{correlated.signature}{The name of the correlated signature
whose exposure is influenced by and co-varies with the exposure
of main.signature. In this study, it defaults as "SBS1".}

\item{name.prefix}{Default: "TwoCorreSigsGen"}

\item{sample.number}{The number of synthetic tumors you want to generate.
Default: 500}

\item{main.mean.log}{The mean of log(count(SBS5),base = 10)
Default: 2.5}

\item{main.stdev.log}{The standard deviation of log(count(SBS5),base = 10)
Default: 0.3}

\item{correlated.stdev.log}{The ADDED standard deviation of log(count(SBS1),base = 10).
This parameter is ADDED stdev because based on the mechanism to generate the count,
log10(count(SBS1)) inherently has a stdev = slope * main.stdev.log
Default: 0.4}

\item{slope.linear}{The ratio for: (Correlated exposure) / (Main exposure) IN LINEAR SPACE!
Default: 0.5}

\item{main.signature.lower.thres}{This program will force the exposure count of
main.signature to be greater than this threhold.
Default: 100}

\item{correlated.signature.lower.thres}{This program will force the exposure count of
correlated.signature to be greater than this threhold.
Default: 1}

\item{pearson.r.2.lower.thres}{Lower boundary of Pearson's R^2
(Default: 0.29)}

\item{pearson.r.2.higher.thres}{Upper boundary of Pearson's R^2
(Default: 0.31)}

\item{min.main.to.correlated.ratio.linear}{The lower ratio for count(SBS5) / count(SBS1)
in LINEAR SPACE! (Default: 1/3)}

\item{max.main.to.correlated.ratio.linear}{(Default: Inf)

\strong{Warning} \cr
Exposure generation function will repeat generating exposure counts
using mean and stdev parameters, until the dataset has a Pearson's R^2
which falls between two boundaries of Pearson's R^2.
Below are a group of parameters which have been tested successfully.
If you intend to lower the Pearson's R^2, do remember to increase
the main.stdev.log and correlated.stdev.log.
Otherwise, the exposure generation will keep generating and discarding datasets!}
}
\description{
This function will use SigProfiler-SBS96 mutational signatures
to generate imaginary tumor spectra with mutation burdens only
from SBS1 and SBS5, and mutation burdens of both signatures
are highly correlated.
}
\details{
If you want to customize the dataset's Pearson R^2,
you need to change the standard deviations of two signatures.
i.e., main.stdev.log and correlated.stdev.log.

This function will generate files listed below:

ground.truth.syn.catalog.csv: Generated tumor spectra in
ICAMS SBS96 CSV format.

ground.truth.syn.exposures.csv: Mutation burdens of SBS1 and
SBS5 in generated tumor spectra in ICAMS CSV format.

ground.truth.syn.sigs.csv: Ground-truth SBS1 and SBS5
signatures in ICAMS SBS96 CSV format.

parameters.txt: Parameters used to generate the exposures
and tumor spectra.

scatterplot.pdf: scatterplot illustrating correlation of
exposures of two signatures in generated spectra

seedInUse.txt, RNGInUse.txt: seed and Random Number Generator
used in generation. (For better reproducibility)

sessionInfo.txt: information related to R versions, platforms,
loaded or imported packages, etc. (For better reproducibility)
}
