% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateTotalRandomSynData2.R
\name{ExposureNums2Exposures}
\alias{ExposureNums2Exposures}
\title{Select \code{num.exp} from the elemtns of \code{sig.names}.}
\usage{
ExposureNums2Exposures(target.num.exp, all.sig.names, target.sig.means,
  target.sig.sds)
}
\arguments{
\item{target.num.exp}{Number of signatures to which sample "was" exposed.}

\item{all.sig.names}{Names of all possible signatures.}

\item{target.sig.means}{Target means for creating mutation counts.}

\item{target.sig.sds}{Target standard deviations for creating mutation counts.}
}
\value{
A subset of \code{sig.names} of size \code{num.exp}.
}
\description{
Select \code{num.exp} from the elemtns of \code{sig.names}.
}
