% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SigSimilarity.R
\name{fwriteDataFrame}
\alias{fwriteDataFrame}
\title{Write a data.frame including rownames using data.table::fwrite}
\usage{
fwriteDataFrame(df, file, rowname.name = "mutation.type")
}
\arguments{
\item{df}{The data.frame to write}

\item{file}{Character string specifying the path to the file to create}

\item{rowname.name}{The colname for the first column on disk, which will
contain the rownames}
}
\description{
Write a data.frame including rownames using data.table::fwrite
}
