% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateSynData.R
\name{GenSynCatalogs}
\alias{GenSynCatalogs}
\title{Generate synthetic spectra catalogs given
signature profiles and synthetic exposures.}
\usage{
GenSynCatalogs(signatures, exposures, profile.info = NULL)
}
\arguments{
\item{signatures}{The signature profiles.}

\item{exposures}{The synthetic exposures.}

\item{profile.info}{Ill-defined string for tumor labels}
}
\value{
Spectra catalog as a numeric matrix.
}
\description{
Generate synthetic spectra catalogs given
signature profiles and synthetic exposures.
}
