% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapSPToSASignatureNamesInExposure.R
\name{MapSPToSASignatureNamesInExposure}
\alias{MapSPToSASignatureNamesInExposure}
\title{With the signatures represented in a matrix of exposures, find the nearest
SignatureAnalyzer exposure.}
\usage{
MapSPToSASignatureNamesInExposure(sp.exposures)
}
\arguments{
\item{sp.exposures}{The exposures}
}
\value{
A list with
\emumerate {
\item \item \code{exp2} \code{sp.exposures} with the
rownames(signature names) updated according to the
match

\item \code{sp.to.sa.sig.match}

\item \code{sa.to.sp.sig.match} Best matches in the opposite direction
}
}
\description{
With the signatures represented in a matrix of exposures, find the nearest
SignatureAnalyzer exposure.
}
\details{
IMPORTANT: uses the package globals \code{\link{sa.96.sigs}}
and \code{\link{sp.sigs}}.
}
