% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MutationalSignatures}
\alias{MutationalSignatures}
\alias{sa.96.sigs}
\alias{sa.COMPOSITE.sigs}
\alias{sp.sigs}
\title{Reference mutational signatures from PCAWG7}
\format{Numerical matrix with rows indicating mutation
types and columns indicating signatures.}
\usage{
sa.96.sigs

sa.COMPOSITE.sigs

sp.sigs
}
\description{
Reference mutational signatures from PCAWG7
}
\keyword{datasets}
