% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MutationalSignatures}
\alias{MutationalSignatures}
\alias{sa.96.sigs}
\alias{sa.COMPOSITE.sigs}
\alias{sp.sigs}
\title{Reference mutational signature profiles from PCAWG7.}
\format{Numerical matrix with rows indicating mutation
types and columns indicating signatures.}
\usage{
sa.96.sigs

sa.COMPOSITE.sigs

sp.sigs
}
\description{
Reference mutational signature profiles from PCAWG7.

SignatureAnalyzer mutational signature profiles collapsed from
COMPOSITE to 96-channel SNS signatures.

COMPOSITE mutational signature profiles extrated by
SignatureAnalyzer.

Signatures extracted by SigProfiler.
}
\keyword{datasets}
