% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateSynData.R
\name{OutDir}
\alias{OutDir}
\title{Create file names in a given directory}
\usage{
OutDir(file.name)
}
\arguments{
\item{file.name}{The name of the that will be
prefixed by \code{out.data.dir}.}
}
\value{
\code{file.name} prefixed by \code{out.data.dir}.
}
\description{
The directory is provided by the global
variable \code{out.data.dir},
which \strong{must} be set by the user. If
\code{out.data.dir} is NULL then just return
\code{file.name}.
}
