% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateSynData.R
\name{present.sigs}
\alias{present.sigs}
\title{Decide which signatures will be present in
 the catalogs of synthetic tumors.}
\usage{
present.sigs(num.tumors, prev.present)
}
\arguments{
\item{num.tumors}{Number of tumors to generate}

\item{prev.present}{Vector of prevalences,
each the prevalence of 1 mutational
signature. The names are the names of the
signatures.}
}
\value{
A matrix with one row
per signature and one column per tumor, with
1 in a cell indicated the presence of a signature
and 0 indicating absence.
}
\description{
Decide which signatures will be present in
 the catalogs of synthetic tumors.
}
\keyword{internal}
