% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignatureAnalyzerInteraction.R
\name{RunSignatureAnalyzerOnFile}
\alias{RunSignatureAnalyzerOnFile}
\title{for COMPOSITE signatures we rbind coordinated
"SNV", "DNP", and "INDEL" signatures. See
\code{\link{ReadSASigCOMPOSITE}}.
Function for running SignatureAnalyzer on a file containing
a catalog.}
\usage{
RunSignatureAnalyzerOnFile(input.catalog, read.catalog.function, out.dir,
  write.signature.function, input.exposures = NULL, maxK = 30,
  tol = 1e-07, test.only = FALSE, delete.tmp.files = TRUE)
}
\arguments{
\item{input.catalog}{File containing input catalog.  Columns are
samples (tumors), rows are signatures.  SignatureAnalyzer does
not care about the row names (I think) TODO(Steve): check this.}

\item{read.catalog.function}{Function taking a file path as
its only argument and returning a catalog as a numeric matrix.}

\item{out.dir}{Directory that will be created for the output;
abort if it already exits.  Log files will be in
\code{paste0(out.dir, "/tmp")}.}

\item{write.signature.function}{Function with first argument the
signatures generated by SignatureAnalyzer and second argument
the file to write to.}

\item{input.exposures}{A file with the synthetic exposures used to generate
\code{input.catalot}; if provided here,
this is copied over to the output directory
for downstream analysis.}

\item{maxK}{The maximum number of signatures to consider
extracting.}

\item{tol}{Controls when SignatureAnalyzer will terminate
its search; \code{tol} was 1.e-05 for the PCAWG7 analysis.}

\item{test.only}{If TRUE, only analyze the first 10 columns
read in from \code{input.catalog}.}

\item{delete.tmp.files}{If TRUE delete the many temporary
files generated by SignatureAnalyzer.}
}
\value{
The output of SignatureAnalyzer, invisibly. This is a list
with five elements (named in this code, not by SignatureAnalyzer):
\enumerate{
 \item \code{signatures.W} The signature matrix
 \item \code{exposures.H} The corresponding exposures
 \item \code{likelihood} The likelihood
 \item \code{evidence} -1 * the posterior probability
 \item \code{relevance} One for each column of the \code{signatures.W}
 \item \code{error} A measure of reconstruction error (?)
}
}
\description{
It is necessary to setwd() to a folder containing
SignatureAnalyzer (right now SignatureAnal\strong{zy}er.052418,
note transposed "zy")
and then call
\preformatted{
options( warn = 0 )
here <- getwd()
setwd("SignatureAnalzyer.052418/") # Or the appropriate directory
INPUT <- "INPUT_SignatureAnalyzer/"
source("SignatureAnalyzer.PCAWG.function.R")
setwd(here) # This is necessary because the caller
            # as specified input and output locations
            # realtive to here.
RunSignatureAnalyzerOnSyn(...)
}
}
\details{
Creates several
 files in \code{paste0(out.dir, "/sa.output.rdata")}. These are
 TODO(Steve): list the files
}
