% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateSynData.R
\name{SetNewOutDir}
\alias{SetNewOutDir}
\title{Globally set the location used by \code{OutDir}.}
\usage{
SetNewOutDir(dir, overwrite = FALSE)
}
\arguments{
\item{dir}{The location of the top level directory}

\item{overwrite}{If TRUE allow overwriting of existing directory}
}
\description{
Globally set the location used by \code{OutDir}.
}
