% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SignatureAnalyzerSigs}
\alias{SignatureAnalyzerSigs}
\alias{sa.96.sigs}
\alias{sa.COMPOSITE.sigs}
\title{Signatures from SignatureAnalyzer}
\format{Numerical matrix with rows indicating mutation
types and columns indicating signatures.}
\usage{
sa.96.sigs

sa.COMPOSITE.sigs
}
\description{
Signatures from SignatureAnalyzer
}
\keyword{datasets}
