% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummarizeResults.R
\name{SignatureAnalyzerSummarizeSBS1SBS5}
\alias{SignatureAnalyzerSummarizeSBS1SBS5}
\title{Summarize all subdirectories of the correlated SBS1 / SBS5.}
\usage{
SignatureAnalyzerSummarizeSBS1SBS5(top.level.dir, overwrite = FALSE)
}
\arguments{
\item{top.level.dir}{Path to top level directory.}

\item{overwrite}{If TRUE overwrite existing directories and files.}
}
\description{
Summarize all subdirectories of the correlated SBS1 / SBS5.
}
