% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignatureAnalyzerInteraction.R
\name{SignatureAnalyzerSummarizeTopLevel}
\alias{SignatureAnalyzerSummarizeTopLevel}
\title{Summarize all subdirs of a major dataset}
\usage{
SignatureAnalyzerSummarizeTopLevel(top.level.dir)
}
\arguments{
\item{top.level.dir}{Path to top level directory}
}
\description{
Summarize all subdirs of a major dataset
}
