% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummarizeResults.R
\name{SignatureAnalyzerSummarizeTopLevel}
\alias{SignatureAnalyzerSummarizeTopLevel}
\title{Summarize all subdirectories of SignatureAnalyzer results
on a major dataset.}
\usage{
SignatureAnalyzerSummarizeTopLevel(top.level.dir, overwrite = FALSE)
}
\arguments{
\item{top.level.dir}{Path to top level directory.}

\item{overwrite}{If TRUE overwrite existing summary files.}
}
\description{
Summarize all subdirectories of SignatureAnalyzer results
on a major dataset.
}
