% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummarizeResults.R
\name{SummarizeSigOneSubdir}
\alias{SummarizeSigOneSubdir}
\title{Assess/evaluate results from SigProfiler or SignatureAnalyzer}
\usage{
SummarizeSigOneSubdir(third.level.dir, ground.truth.exposure.name,
  extracted.sigs.path, read.extracted.sigs.fn, read.ground.truth.sigs.fn,
  write.png = FALSE)
}
\arguments{
\item{third.level.dir}{Lowest level path to results, e.g.
\code{top.dir}/sa.sa.96/sp.results/. or
\code{top.dir}/sa.sa.96/sa.results/
Here, \code{top.dir} refers to a top-level directory which contains the
full information of a synthetic dataset. (e.g. \code{syn.2.7a.7b.abst.v8})
This code depends on a conventional directory structure documented
elsewhere. However there should be a directory within the \code{third.level.dir}
which stores the software output.}

\item{ground.truth.exposure.name}{File name which stores ground-truth exposures.
It defaults to be "ground.truth.syn.exposures.csv". This file can be found
in the \code{sub.dir}, i.e. \code{third.level.dir}/../}

\item{extracted.sigs.path}{Full path of extracted sigs file, e.g.
\code{<third.level.dir>/SBS96/Selected_Solution/De_Novo_Solution/signatures.PCAWG.format.csv}.}

\item{read.extracted.sigs.fn}{Function to read the extracted sigs file.
e.g. \code{ReadCat96}}

\item{read.ground.truth.sigs.fn}{Function to read the ground-truth sigs file.
e.g. \code{ReadCat96}}

\item{write.png}{If TRUE create png plots of the signatures.}
}
\description{
Assess/evaluate results from SigProfiler or SignatureAnalyzer
}
