% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EvaluateResults.R
\name{SummarizeSigOneSubdir}
\alias{SummarizeSigOneSubdir}
\title{Assess/evaluate results from SigProfiler or SignatureAnalyzer}
\usage{
SummarizeSigOneSubdir(path, ground.truth.exposures, ground.truth.sigs,
  write.png = FALSE)
}
\arguments{
\item{path}{Lowest level path to results, e.g.
ATopLevelDir/sa.sa.96/sp.results/.
This code depends on a conventional directory structure documented
elsewhere. However there should be a directory <path>/SBS96 which
stores SigProfiler results.}

\item{ground.truth.exposures}{One of "sa.exposure.csv" or "sp.exposure.csv".}

\item{ground.truth.sigs}{One of \code{sp.sigs} or \code{sa.96.sigs}.}

\item{write.png}{If TRUE create png plots of the signatures.}
}
\description{
Assess/evaluate results from SigProfiler or SignatureAnalyzer
}
