% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummarizeResults.R
\name{SummarizeSigOneSubdir}
\alias{SummarizeSigOneSubdir}
\title{Assess/evaluate results from Multiple software packages}
\usage{
SummarizeSigOneSubdir(third.level.dir, ground.truth.exposure.name,
  extracted.sigs.path, attributed.exp.path = NULL,
  read.extracted.sigs.fn, read.ground.truth.sigs.fn, write.cat.fn,
  plot.pdf.fn, overwrite = FALSE, summary.folder.name = "summary")
}
\arguments{
\item{third.level.dir}{Lowest level path to results, that is,
\code{top.dir}/sp.sp/sa.results/. or
\code{top.dir}/sa.sa.96/sa.results/
Here, \code{top.dir} refers to a top-level directory which contains the
full information of a synthetic dataset. (e.g. \code{syn.2.7a.7b.abst.v8})
This code depends on a conventional directory structure documented
elsewhere. However there should be a directory within the \code{third.level.dir}
which stores the software output.}

\item{ground.truth.exposure.name}{File name which stores ground-truth exposures.
It defaults to be "ground.truth.syn.exposures.csv". This file can be found
in the \code{sub.dir}, i.e. \code{third.level.dir}/../}

\item{extracted.sigs.path}{Path to extracted sigs file, e.g.
\code{<third.level.dir>/SBS96/Selected_Solution/De_Novo_Solution/signatures.PCAWG.format.csv}.}

\item{attributed.exp.path}{Path to attributed exposures file.}

\item{read.extracted.sigs.fn}{Function to read the extracted sigs file.
e.g. \code{ReadCatSNS96}}

\item{read.ground.truth.sigs.fn}{Function to read the ground-truth sigs file.
e.g. \code{ReadCatSNS96}}

\item{write.cat.fn}{Function to write a catalog to disk, for example
[ICAMS]\code{WriteCatSNS96}}

\item{plot.pdf.fn}{If a function, use it to plot PDFs of the ground truth and
extracted signatures.}

\item{overwrite}{If TRUE overwrite existing directories and files.}

\item{summary.folder.name}{The name of the folder containing summary results.
Usually, it equals to "summary", but it shall be changed in case a software
(e.g. SignatureEstimation) has results from 2 or more different methods
(e.g. QP - Quadratic Programming; SA - Simulated Annealing). In this case,
the summary folder name shall be different (e.g. "summary.QP" and "summary.SA").
summary folder will be created under the \code{third.level.dir}.}
}
\description{
Note: Users should use sigproextractor(SigProfiler-Python) v0.0.5.43
and SignatureAnalyzer 2018-Apr-18
}
\keyword{internal}
