% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EvaluateResults.R, R/SummarizeResults.R
\name{SummarizeSigOneSubdir}
\alias{SummarizeSigOneSubdir}
\title{Assess/evaluate results from SigProfiler or SignatureAnalyzer}
\usage{
SummarizeSigOneSubdir(third.level.dir, ground.truth.exposure.name,
  extracted.sigs.path, read.extracted.sigs.fn, read.ground.truth.sigs.fn,
  write.png = FALSE)

SummarizeSigOneSubdir(third.level.dir, ground.truth.exposure.name,
  extracted.sigs.path, read.extracted.sigs.fn, read.ground.truth.sigs.fn,
  write.png = FALSE)
}
\arguments{
\item{third.level.dir}{Lowest level path to results, e.g.
ATopLevelDir/sa.sa.96/sp.results/.
This code depends on a conventional directory structure documented
elsewhere. However there should be a directory <third.level.dir>/SBS96 which
stores SigProfiler results.}

\item{ground.truth.exposure.name}{One of "sa.exposure.csv" or "sp.exposure.csv".}

\item{extracted.sigs.path}{XXXXX}

\item{read.extracted.sigs.fn}{XXXXXX}

\item{read.ground.truth.sigs.fn}{XXXXXXX}

\item{write.png}{If TRUE create png plots of the signatures.}

\item{path}{Lowest level path to results, e.g.
ATopLevelDir/sa.sa.96/sp.results/.
This code depends on a conventional directory structure documented
elsewhere. However there should be a directory <path>/SBS96 which
stores SigProfiler results.}

\item{software.name}{Name of the software from which the results are generated.
It can be the full name of the software (e.g. One in \code{c("sigproextractor","SigProfiler","SignatureAnalyzer")});
it can also be the acronym of the software (e.g. \code{c("sp","sa")} is short for
\code{c("SigProfiler","SignatureAnalyzer")}.}

\item{ground.truth.exposures}{One of "sa.exposure.csv" or "sp.exposure.csv".}

\item{ground.truth.sigs}{One of \code{sp.sigs} or \code{sa.96.sigs}.}

\item{write.png}{If TRUE create png plots of the signatures.}
}
\description{
Assess/evaluate results from SigProfiler or SignatureAnalyzer

Assess/evaluate results from SigProfiler or SignatureAnalyzer
}
