% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EvaluateResults.R, R/SummarizeResults.R
\name{SummarizeSigProfiler}
\alias{SummarizeSigProfiler}
\title{Summarize SigProfiler results in the sa.sa.96 and sp.sp subdirectories}
\usage{
SummarizeSigProfiler(top.dir, write.png = FALSE)

SummarizeSigProfiler(top.dir, write.png = FALSE)
}
\arguments{
\item{top.dir}{The top directory of a conventional data structure containing
at least one of the subdirectories: sa.sa.96/sp.results and sp.sp/sp.results;
see further documentation elsewhere.}

\item{write.png}{If TRUE create png plots of the signatures.}

\item{sub.dir}{The subdirectory under \code{top.dir}, and containing a folder
named sp.results. By default, it equals to \code{c("sa.sa.96","sp.sp")}. But
you should specify \code{sub.dir = "sp.sp"} for clock-like SBS1-and-SBS5-containing
datasets.}

\item{top.dir}{The top directory of a conventional data structure containing
subdirectories sa.sa.96/sp.results and sp.sp/sp.results; see further
documentation elsewhere.}

\item{write.png}{If TRUE create png plots of the signatures.}
}
\description{
Summarize SigProfiler results in the sa.sa.96 and sp.sp subdirectories

Summarize SigProfiler results in the sa.sa.96 and sp.sp subdirectories
}
\details{
Results are put in standardized subdirectories of \code{top.dir}.

Results are put in standardized subdirectories of \code{top.dir}.
}
