% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EvaluateResults.R
\name{SummarizeSigProfiler}
\alias{SummarizeSigProfiler}
\title{Summarize SigProfiler results in the sa.sa.96 and sp.sp subdirectories}
\usage{
SummarizeSigProfiler(top.dir, write.png = FALSE)
}
\arguments{
\item{top.dir}{The top directory of a conventional data structure containing
subdirectories sa.sa.96/sp.results and sp.sp/sp.results; see further
documentation elsewhere.}

\item{write.png}{If TRUE create png plots of the signatures.}
}
\description{
Summarize SigProfiler results in the sa.sa.96 and sp.sp subdirectories
}
\details{
Results are put in standardized subdirectories of \code{top.dir}.
}
