% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateSynData.R
\name{synsig.params.from.attributions}
\alias{synsig.params.from.attributions}
\title{Determine 3 parameters for synthetic tumors from an exposure matrix}
\usage{
synsig.params.from.attributions(counts, target.size = 1)
}
\arguments{
\item{counts}{A matrix in which each column is a sample and each row is a mutation
signature, with each element being the "exposure",
i.e. mutation count attributed to a
(sample, signature) pair.}

\item{target.size}{Deprecated - was the length of sequence
from which the counts were derived; in the future
make any adjustments (e.g exome to genome) before
providing the exposure matrix.}
}
\value{
A data frame with one row for
each of a subset of the input signatures
 and the following columns. Signatures not present in
 \code{counts} or present only in a single tumor in \code{counts} are removed.

\enumerate{
\item the proportion of tumors with the signature
\item mean(log_10(mutations.per.Mb))
\item stddev(log_10(mutations.per.Mb))
}
}
\description{
Determine 3 parameters for synthetic tumors from an exposure matrix
}
