% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SynSig.R
\docType{package}
\name{SynSig}
\alias{SynSig}
\alias{SynSig-package}
\title{SynSig: A package for working with synthetic mutational signatures for
purpose of evaluating software packages for extracting and attributing
mutational signatures.}
\description{
The SynSig package provides several categories
of functions for working with mutational signatures.
}
\section{Creating Synthetic Mutational Catalogs}{


These functions create synthetic mutational catalogs based
on parameters derived from signature profiles
and attributions (exposures).
}

\section{Signature Matching and Comparison}{


Functions for comparing mutational signatures and
sets of mutational signatures:
\code{\link{Match1Sig}},
\code{\link{MatchSigs1Direction}},
\code{\link{MatchSigs2Directions}},
\code{\link{MatchSigsThenWriteAndPlot}},
\code{\link{TestMatchSigs2Directions}}
}

\section{Miscellaneous}{

fwriteDataFrame
}

\section{Overview}{


Generate synthetic mutational spectra (mutations in tumors) from
\enumerate{
\item A set of mutational signatures

\item Attribution of mutation counts in each tumor to the responsible mutational
signatures
}


Typical workflow

\preformatted{
A <- Read in matrix of attributions (signatures x samples)
S <- Read the mutational signature profiles.

P <- synsig.parameters.from.attributions (A, ....)

No.noise.exposure <- target.size * generate.synthetic.exposures (P....)

No.noise.specra <- GenSynCatalogs(S, No.noise.exposure)

# Skip for now
# Noise.spectra <- ... neg binomial resample from No.noise.spectra

Use functions from ICAMS to save spectra

T <-  Results of running extraction package on spectra

MatchSigsThenWriteAndPlot(T, S, ...)

}

For more information, refer to

https://docs.google.com/document/d/183DzzG80BxuAdcnuL1jD1qI_YM-cgtVEdHvkbMLwHGc/edit?usp=sharing

TODO(steve): describe how to evaluate!
}

\keyword{internal}
