% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateSynData.R
\name{SynSigParamsOneSignature}
\alias{SynSigParamsOneSignature}
\title{Extract SynSig parameters for one mutational signature profile}
\usage{
SynSigParamsOneSignature(counts, target.size)
}
\arguments{
\item{counts}{A vector of mutation counts attributed to one signature across
length(counts) samples. TODO(Steve): rename to exposures}

\item{target.size}{The length of genomic sequence from which the counts
were derived, in megabases TODO:(Steve) probably do not need this}
}
\value{
A 3-element vector with names "prevalence", "mean", and "stdev"
}
\description{
Extract SynSig parameters for one mutational signature profile
}
