% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadWriteCatalogs.R
\name{WriteCatCOMPOSITE}
\alias{WriteCatCOMPOSITE}
\title{Write a COMPOSITE catalog or signature matrix to disk}
\usage{
WriteCatCOMPOSITE(ct, path)
}
\arguments{
\item{ct}{A catalog or signature matrix}

\item{path}{Path to file to write}
}
\description{
Write a COMPOSITE catalog or signature matrix to disk
}
\seealso{
\code{\link{ReadCatCOMPOSITE}}
}
